% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{as_latex}
\alias{as_latex}
\title{Distribution as LaTeX expression}
\usage{
as_latex(distribution, digits = 3)
}
\arguments{
\item{distribution}{A \code{\link[=Distribution]{Distribution()}} object}

\item{digits}{Integer to which to round digits}
}
\description{
Get a summary of a distribution as a LaTeX expression
}
