% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{bc_transform}
\alias{bc_transform}
\title{Box-Cox Transformation}
\usage{
bc_transform(lambda)
}
\arguments{
\item{lambda}{Numeric. The transformation parameter.}
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for the one-parameter Box-Cox transformation
\eqn{Y = (X^\lambda - 1)/\lambda} (with \eqn{Y = \log X} for \eqn{\lambda = 0}).
Requires \eqn{X > 0}.
}
\examples{
d <- transformation(gamma2_distrib(), bc_transform(lambda = 0.5))
distrib_pdf(d, 1, list(mu = 2, sigma2 = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{logit_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
