% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-chi_squared.R
\name{chi_squared}
\alias{chi_squared}
\alias{noncentral_chi_squared}
\alias{ChiSquared}
\title{Chi-squared distribution}
\usage{
chi_squared(nu)

noncentral_chi_squared(nu, kappa)

ChiSquared(nu, kappa)
}
\arguments{
\item{nu}{degrees of freedom paramter.}

\item{kappa}{noncentrality parameter.}
}
\description{
Create a chi-squared distribution object.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
