% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{weights_ic}
\title{Compare distributions}
\usage{
compare(..., data)

weights_ic(x)
}
\arguments{
\item{...}{Objects of class \code{\link[=Distribution]{Distribution()}}.}

\item{data}{Numeric containing the data.}

\item{x}{Numberic vector of AIC/BIC values.}
}
\value{
A data.frame containing the results.
}
\description{
Compare distributions using information criteria.
}
\details{
Computes \code{\link[=information_criteria]{information_criteria()}} for each distribution.
Note that prior to comparing the distributions, you ought to fit the distributions with \code{\link[=Mle]{Mle()}} (see \code{\link[=fit_distribution]{fit_distribution()}}).
}
