% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_cdf.Beta2Distrib}
\alias{distrib_cdf.Beta2Distrib}
\title{Beta Distribution Function in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The incomplete beta function.
}
\seealso{
\code{\link[=beta2_distrib]{beta2_distrib()}}
}
