% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded.R
\name{distrib_cdf.FoldedDistrib}
\alias{distrib_cdf.FoldedDistrib}
\title{Folded Distribution Function}
\arguments{
\item{distrib}{A \code{FoldedDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A named list of the parent's parameters.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \eqn{P(|Y| \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities are returned as logarithms.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
\deqn{P(|Y| \le q) = F(q) - F(-q)}
}
\seealso{
\code{\link[=folded]{folded()}}
}
