% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_distrib.R
\name{distrib_cdf.GPDDistrib}
\alias{distrib_cdf.GPDDistrib}
\title{Generalized Pareto Distribution Function}
\arguments{
\item{distrib}{A \code{GPDDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing \code{sigma} and \code{xi}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities are returned as logarithms.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
\deqn{F(q) = 1 - \left(1 + \dfrac{\xi q}{\sigma}\right)^{-1/\xi}}
}
\seealso{
\code{\link[=gpd_distrib]{gpd_distrib()}}
}
