% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_distrib.R
\name{distrib_expected_hessian.GaussianDistrib}
\alias{distrib_expected_hessian.GaussianDistrib}
\title{Gaussian Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{GaussianDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Gaussian log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\dfrac{1}{\sigma^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \sigma^2}\right] = -\dfrac{2}{\sigma^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma}\right] = 0}
}
\seealso{
\code{\link{gaussian_distrib}}
}
