% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss_distrib.R
\name{distrib_expected_hessian.InvGaussDistrib}
\alias{distrib_expected_hessian.InvGaussDistrib}
\title{Inverse-Gaussian Analytical Expected Hessian}
\arguments{
\item{distrib}{An \code{InvGaussDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Inverse-Gaussian log-density
with respect to the parameters \eqn{\mu} and \eqn{\phi}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\dfrac{1}{\phi\mu^3}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \phi^2}\right] = -\dfrac{1}{2\phi^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu \partial \phi}\right] = 0}
}
\seealso{
\code{\link{invgauss_distrib}}
}
