% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_grad_y.Beta2Distrib}
\alias{distrib_grad_y.Beta2Distrib}
\title{Beta Response Derivatives in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial\ell/\partial y = (\alpha-1)/y - (\beta-1)/(1-y)}.
}
\seealso{
\code{\link{beta2_distrib}}
}
