% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet_distrib.R
\name{distrib_grad_y.EnetDistrib}
\alias{distrib_grad_y.EnetDistrib}
\title{Elastic-Net Response Derivatives}
\arguments{
\item{distrib}{An \code{EnetDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{mu}, \code{lambda} and \code{alpha}.}
}
\value{
A numeric vector.
}
\description{
\eqn{\ell^{(y)} = -a\,\mathrm{sgn}(y-\mu) - c(y-\mu)},
undefined at the location.
}
\seealso{
\code{\link{enet_distrib}}
}
