% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distrib.R
\name{distrib_grad_y.GammaDistrib}
\alias{distrib_grad_y.GammaDistrib}
\title{Gamma Response Derivatives}
\arguments{
\item{distrib}{A \code{GammaDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector.
}
\description{
Closed-form derivatives of the Gamma log-density with respect to the response,
using the shape/rate reparameterization \eqn{\alpha = \mu^2/\sigma^2},
\eqn{\lambda = \mu/\sigma^2}:
\eqn{\partial \ell / \partial y = (\alpha-1)/y - \lambda} and
\eqn{\partial^2 \ell / \partial y^2 = -(\alpha-1)/y^2}.
}
\seealso{
\code{\link{gamma_distrib}}
}
