% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_distrib.R
\name{distrib_grad_y.GaussianDistrib}
\alias{distrib_grad_y.GaussianDistrib}
\title{Gaussian Response Derivatives}
\arguments{
\item{distrib}{A \code{GaussianDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector.
}
\description{
Closed-form derivatives of the Gaussian log-density with respect to the response:
\eqn{\partial \ell / \partial y = -(y-\mu)/\sigma^2} and
\eqn{\partial^2 \ell / \partial y^2 = -1/\sigma^2}.
}
\seealso{
\code{\link{gaussian_distrib}}
}
