% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmises_distrib.R
\name{distrib_grad_y.VonMisesDistrib}
\alias{distrib_grad_y.VonMisesDistrib}
\title{von Mises Response Gradient}
\arguments{
\item{distrib}{A \code{VonMisesDistrib} object.}

\item{y}{A numeric vector of angles.}

\item{theta}{A list containing \code{mu} and \code{kappa}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\deqn{\dfrac{\partial \ell}{\partial y} = -\kappa \sin(y - \mu)}
}
\seealso{
\code{\link{vonmises_distrib}}
}
