% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distrib.R
\name{distrib_grad_y.WeibullDistrib}
\alias{distrib_grad_y.WeibullDistrib}
\title{Weibull Response Derivative}
\arguments{
\item{distrib}{A \code{WeibullDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\partial \ell / \partial y = (\sigma - 1 - \sigma u)/y}.
}
\seealso{
\code{\link{weibull_distrib}}
}
