% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distrib.R
\name{distrib_gradient.BernoulliDistrib}
\alias{distrib_gradient.BernoulliDistrib}
\title{Bernoulli Analytical Gradient}
\arguments{
\item{distrib}{A \code{BernoulliDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of first derivatives.
}
\description{
Computes the analytical gradient (first derivative) of the Bernoulli log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{y - \mu}{\mu(1-\mu)}}
}
\seealso{
\code{\link[=bernoulli_distrib]{bernoulli_distrib()}}
}
