% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distrib.R
\name{distrib_gradient.GammaDistrib}
\alias{distrib_gradient.GammaDistrib}
\title{Gamma Analytical Gradient}
\arguments{
\item{distrib}{A \code{GammaDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Gamma log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma^2}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{-2\mu\psi\left(\dfrac{\mu^2}{\sigma^2}\right) + 2\mu\log\left(\dfrac{\mu}{\sigma^2}\right) + \mu + 2\mu\log(y) - y}{\sigma^2}}
\deqn{\dfrac{\partial \ell}{\partial \sigma^2} = -\dfrac{\mu\left[-\mu\psi\left(\dfrac{\mu^2}{\sigma^2}\right) + \mu + \mu\left(\log\left(\dfrac{\mu}{\sigma^2}\right) + \log(y)\right) - y\right]}{(\sigma^2)^2}}
}
\seealso{
\code{\link{gamma_distrib}}
}
