% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian1_distrib.R
\name{distrib_gradient.Gaussian1Distrib}
\alias{distrib_gradient.Gaussian1Distrib}
\title{Gaussian Analytical Gradient}
\arguments{
\item{distrib}{A \code{Gaussian1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Gaussian log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{y - \mu}{\sigma^2}}
\deqn{\dfrac{\partial \ell}{\partial \sigma} = \dfrac{(y - \mu)^2 - \sigma^2}{\sigma^3}}
}
\seealso{
\code{\link{gaussian1_distrib}}
}
