% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss_distrib.R
\name{distrib_gradient.InvGaussDistrib}
\alias{distrib_gradient.InvGaussDistrib}
\title{Inverse-Gaussian Analytical Gradient}
\arguments{
\item{distrib}{An \code{InvGaussDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Inverse-Gaussian log-density
with respect to the parameters \eqn{\mu} and \eqn{\phi}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{y - \mu}{\phi\mu^3}}
\deqn{\dfrac{\partial \ell}{\partial \phi} = \dfrac{(y - \mu)^2 - y\mu^2\phi}{2y\phi^2\mu^2}}
}
\seealso{
\code{\link{invgauss_distrib}}
}
