% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distrib.R
\name{distrib_gradient.LogisticDistrib}
\alias{distrib_gradient.LogisticDistrib}
\title{Logistic Analytical Gradient}
\arguments{
\item{distrib}{A \code{LogisticDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Logistic log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{1}{\sigma} \tanh\left(\dfrac{y-\mu}{2\sigma}\right)}
\deqn{\dfrac{\partial \ell}{\partial \sigma} = -\dfrac{1}{\sigma} \left[ 1 - \dfrac{y-\mu}{\sigma} \tanh\left(\dfrac{y-\mu}{2\sigma}\right) \right]}
}
\seealso{
\code{\link[=logistic_distrib]{logistic_distrib()}}
}
