% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distrib.R
\name{distrib_gradient.LognormalDistrib}
\alias{distrib_gradient.LognormalDistrib}
\title{Lognormal Analytical Gradient}
\arguments{
\item{distrib}{A \code{LognormalDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Lognormal log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma^2}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{\log(y) - \mu}{\sigma^2}}
\deqn{\dfrac{\partial \ell}{\partial \sigma^2} = \dfrac{(\log(y) - \mu)^2 - \sigma^2}{2(\sigma^2)^2}}
}
\seealso{
\code{\link{lognormal_distrib}}
}
