% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_distrib.R
\name{distrib_gradient.PseudoHuberDistrib}
\alias{distrib_gradient.PseudoHuberDistrib}
\title{Pseudo-Huber Analytical Gradient}
\arguments{
\item{distrib}{A \code{PseudoHuberDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient of the Pseudo-Huber log-density.
Let \eqn{r = y - \mu} and \eqn{D = \sqrt{\nu + (r/\sigma)^2}}:

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{r}{\sigma^2 D}}
\deqn{\dfrac{\partial \ell}{\partial \sigma} = \dfrac{1}{\sigma} \left( \dfrac{r^2}{\sigma^2 D} - 1 \right)}
\deqn{\dfrac{\partial \ell}{\partial \nu} = -\dfrac{1}{2} \left[ \dfrac{1}{\nu} + \dfrac{1}{D} + \dfrac{K_1'(\sqrt{\nu})}{\sqrt{\nu}\, K_1(\sqrt{\nu})} \right]}
}
\seealso{
\code{\link{pseudohuber_distrib}}
}
