% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta1_distrib.R
\name{distrib_hess_y.Beta1Distrib}
\alias{distrib_hess_y.Beta1Distrib}
\title{Beta Response Second Derivative}
\arguments{
\item{distrib}{A \code{Beta1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = -(\alpha-1)/y^2 - (\beta-1)/(1-y)^2}.
}
\seealso{
\code{\link{beta1_distrib}}
}
