% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_distrib.R
\name{distrib_hess_y.ChisqDistrib}
\alias{distrib_hess_y.ChisqDistrib}
\title{Chi-Squared Response Hessian}
\arguments{
\item{distrib}{A \code{ChisqDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\deqn{\dfrac{\partial^2 \ell}{\partial y^2} = -\dfrac{\mu/2 - 1}{y^2}}
}
\seealso{
\code{\link[=chisq_distrib]{chisq_distrib()}}
}
