% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_hess_y.ExponentialDistrib}
\alias{distrib_hess_y.ExponentialDistrib}
\title{Exponential Response Hessian}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{...}{Unused.}
}
\value{
A numeric vector of zeros.
}
\description{
\deqn{\dfrac{\partial^2 \ell}{\partial y^2} = 0}
the log-density being linear in the response.
}
\seealso{
\code{\link[=exponential_distrib]{exponential_distrib()}}
}
