% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_distrib.R
\name{distrib_hess_y.Gamma1Distrib}
\alias{distrib_hess_y.Gamma1Distrib}
\title{Gamma Second Response Derivative in Mean and Dispersion}
\arguments{
\item{distrib}{A \code{Gamma1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{phi}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial^2\ell/\partial y^2 = -(a-1)/y^2}.
}
\seealso{
\code{\link[=gamma1_distrib]{gamma1_distrib()}}
}
