% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distrib.R
\name{distrib_hess_y.GammaDistrib}
\alias{distrib_hess_y.GammaDistrib}
\title{Gamma Response Second Derivative}
\arguments{
\item{distrib}{A \code{GammaDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = -(\alpha-1)/y^2}, \eqn{\alpha = \mu^2/\sigma^2}.
}
\seealso{
\code{\link{gamma_distrib}}
}
