% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian1_distrib.R
\name{distrib_hess_y.Gaussian1Distrib}
\alias{distrib_hess_y.Gaussian1Distrib}
\title{Gaussian Response Second Derivative}
\arguments{
\item{distrib}{A \code{Gaussian1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = -1/\sigma^2}.
}
\seealso{
\code{\link[=gaussian1_distrib]{gaussian1_distrib()}}
}
