% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss_distrib.R
\name{distrib_hess_y.InvGaussDistrib}
\alias{distrib_hess_y.InvGaussDistrib}
\title{Inverse-Gaussian Response Second Derivative}
\arguments{
\item{distrib}{An \code{InvGaussDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = 3/(2y^2) - 1/(\phi y^3)}.
}
\seealso{
\code{\link{invgauss_distrib}}
}
