% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distrib.R
\name{distrib_hess_y.LaplaceDistrib}
\alias{distrib_hess_y.LaplaceDistrib}
\title{Laplace Response Second Derivative}
\arguments{
\item{distrib}{A \code{LaplaceDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{b}.}
}
\value{
A numeric vector of zeros.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = 0} (almost everywhere).
}
\seealso{
\code{\link{laplace_distrib}}
}
