% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distrib.R
\name{distrib_hess_y.LognormalDistrib}
\alias{distrib_hess_y.LognormalDistrib}
\title{Lognormal Response Second Derivative}
\arguments{
\item{distrib}{A \code{LognormalDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = (1 + (\log y - \mu - 1)/\sigma^2)/y^2}.
}
\seealso{
\code{\link{lognormal_distrib}}
}
