% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmises1_distrib.R
\name{distrib_hess_y.VonMises1Distrib}
\alias{distrib_hess_y.VonMises1Distrib}
\title{von Mises Response Hessian}
\arguments{
\item{distrib}{A \code{VonMises1Distrib} object.}

\item{y}{A numeric vector of angles.}

\item{theta}{A list containing \code{mu} and \code{kappa}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\deqn{\dfrac{\partial^2 \ell}{\partial y^2} = -\kappa \cos(y - \mu)}
}
\seealso{
\code{\link{vonmises1_distrib}}
}
