% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distrib.R
\name{distrib_hessian.BetaDistrib}
\alias{distrib_hessian.BetaDistrib}
\title{Beta Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{BetaDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivatives) of the Beta log-density
with respect to the parameters \eqn{\mu} and \eqn{\phi}.
}
\seealso{
\code{\link{beta_distrib}}
}
