% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distrib.R
\name{distrib_hessian.BinomialDistrib}
\alias{distrib_hessian.BinomialDistrib}
\title{Binomial Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{BinomialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivative) of the Binomial log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = -\dfrac{y}{\mu^2} - \dfrac{n-y}{(1-\mu)^2}}
}
\seealso{
\code{\link[=binomial_distrib]{binomial_distrib()}}
}
