% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin2_distrib.R
\name{distrib_hessian.NegBin2Distrib}
\alias{distrib_hessian.NegBin2Distrib}
\title{Negative Binomial Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{NegBin2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{theta}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivatives) of the Negative Binomial log-probability
with respect to the parameters \eqn{\mu} and \eqn{\theta}.

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = \dfrac{y+\theta}{(\theta+\mu)^2} - \dfrac{y}{\mu^2}}
\deqn{\dfrac{\partial^2 \ell}{\partial \theta^2} = \psi_1(y+\theta) - \psi_1(\theta) + \dfrac{\mu}{\theta(\theta+\mu)} + \dfrac{y-\mu}{(\theta+\mu)^2}}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \partial \theta} = \dfrac{y-\mu}{(\theta+\mu)^2}}
}
\seealso{
\code{\link{negbin2_distrib}}
}
