% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distrib.R
\name{distrib_hessian.WeibullDistrib}
\alias{distrib_hessian.WeibullDistrib}
\title{Weibull Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{WeibullDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{scale}{Either \code{"parameter"} or \code{"link"}.}

\item{...}{Unused.}
}
\value{
A named list of second derivatives.
}
\description{
Closed-form second derivatives of the Weibull log-density:
\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2}
        = \dfrac{\sigma}{\mu^2}\left\{1 - (1 + \sigma) u\right\},
      \qquad
      \dfrac{\partial^2 \ell}{\partial \sigma^2}
        = -\dfrac{1}{\sigma^2} - u (\log z)^2,}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \, \partial \sigma}
        = \dfrac{1}{\mu}\left(u - 1 + \sigma u \log z\right).}
}
\seealso{
\code{\link{weibull_distrib}}
}
