% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_pdf.ExponentialDistrib}
\alias{distrib_pdf.ExponentialDistrib}
\title{Exponential Density}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector of density values.
}
\description{
\deqn{f(y; \mu) = \dfrac{1}{\mu} e^{-y/\mu}, \qquad y > 0}
}
\seealso{
\code{\link{exponential_distrib}}
}
