% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_distrib.R
\name{distrib_pdf.Gamma1Distrib}
\alias{distrib_pdf.Gamma1Distrib}
\title{Gamma Density in Mean and Dispersion}
\arguments{
\item{distrib}{A \code{Gamma1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{phi}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector.
}
\description{
The gamma density at shape \eqn{a = 1/\phi} and rate
\eqn{b = 1/(\phi\mu)}.
}
\seealso{
\code{\link[=gamma1_distrib]{gamma1_distrib()}}
}
