% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distrib.R
\name{distrib_pdf.GeometricDistrib}
\alias{distrib_pdf.GeometricDistrib}
\title{Geometric Probability Mass Function}
\arguments{
\item{distrib}{A \code{GeometricDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{log}{Logical; if \code{TRUE}, returns the log-probability.}
}
\value{
A numeric vector of probability values.
}
\description{
\deqn{P(Y = y; \mu) = \dfrac{1}{1+\mu}
      \left(\dfrac{\mu}{1+\mu}\right)^{y}, \qquad y = 0, 1, 2, \dots}
}
\seealso{
\code{\link{geometric_distrib}}
}
