% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_distrib.R
\name{distrib_pdf.PseudoHuberDistrib}
\alias{distrib_pdf.PseudoHuberDistrib}
\title{Pseudo-Huber Probability Density Function}
\arguments{
\item{distrib}{A \code{PseudoHuberDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector of density values.
}
\description{
Computes the probability density function for the Pseudo-Huber distribution:
\deqn{f(y; \mu, \sigma, \nu) = \dfrac{1}{2 \sigma \sqrt{\nu} K_1(\sqrt{\nu})} \exp\left( - \sqrt{\nu + \left(\dfrac{y-\mu}{\sigma}\right)^2} \right)}
where \eqn{K_1} is the modified Bessel function of the second kind.
}
\seealso{
\code{\link{pseudohuber_distrib}}
}
