% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_quantile.ExponentialDistrib}
\alias{distrib_quantile.ExponentialDistrib}
\title{Exponential Quantile Function}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \code{p} is \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is given as its logarithm.}
}
\value{
A numeric vector of quantiles.
}
\description{
\deqn{Q(p; \mu) = -\mu \log(1 - p)}
}
\seealso{
\code{\link{exponential_distrib}}
}
