% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss2_distrib.R
\name{distrib_quantile.InvGauss2Distrib}
\alias{distrib_quantile.InvGauss2Distrib}
\title{Inverse Gaussian Quantile Function in Mean and Shape}
\arguments{
\item{distrib}{An \code{InvGauss2Distrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list with \code{mu} and \code{lambda}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is a log-probability.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The inverse gaussian quantile function at dispersion
\eqn{1/\lambda}.
}
\seealso{
\code{\link[=invgauss2_distrib]{invgauss2_distrib()}}
}
