% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin1_distrib.R
\name{distrib_quantile.NegBin1Distrib}
\alias{distrib_quantile.NegBin1Distrib}
\title{NB1 Quantile Function}
\arguments{
\item{distrib}{A \code{NegBin1Distrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing \code{mu} and \code{theta}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \code{p} is \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is given as its logarithm.}
}
\value{
A numeric vector of quantiles.
}
\description{
\code{\link[stats:NegBinomial]{stats::qnbinom()}} at the same size and probability.
}
\seealso{
\code{\link[=negbin1_distrib]{negbin1_distrib()}}
}
