% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom2_distrib.R
\name{distrib_rng.BetaBinom2Distrib}
\alias{distrib_rng.BetaBinom2Distrib}
\title{Beta-Binomial Random Generation in Its Shapes}
\arguments{
\item{distrib}{A \code{BetaBinom2Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{alpha} and \code{beta}.}
}
\value{
A numeric vector of counts.
}
\description{
A beta draw for the probability, then a binomial draw.
}
\seealso{
\code{\link{betabinom2_distrib}}
}
