% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distrib.R
\name{distrib_rng.BetaDistrib}
\alias{distrib_rng.BetaDistrib}
\title{Beta Random Number Generator}
\arguments{
\item{distrib}{A \code{BetaDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Beta distribution.
}
\seealso{
\code{\link{beta_distrib}}
}
