% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2_distrib.R
\name{distrib_rng.Gamma2Distrib}
\alias{distrib_rng.Gamma2Distrib}
\title{Gamma Random Number Generator}
\arguments{
\item{distrib}{A \code{Gamma2Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Gamma distribution.
}
\seealso{
\code{\link{gamma2_distrib}}
}
