% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian2_distrib.R
\name{distrib_rng.Gaussian2Distrib}
\alias{distrib_rng.Gaussian2Distrib}
\title{Gaussian Random Generation in Mean and Variance}
\arguments{
\item{distrib}{A \code{Gaussian2Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector.
}
\description{
Delegates to \code{\link[stats:Normal]{stats::rnorm()}}.
}
\seealso{
\code{\link[=gaussian2_distrib]{gaussian2_distrib()}}
}
