% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_distrib.R
\name{distrib_rng.GaussianDistrib}
\alias{distrib_rng.GaussianDistrib}
\title{Gaussian Random Number Generator}
\arguments{
\item{distrib}{A \code{GaussianDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Gaussian distribution.
}
\seealso{
\code{\link{gaussian_distrib}}
}
