% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_distrib.R
\name{distrib_rng.GumbelDistrib}
\alias{distrib_rng.GumbelDistrib}
\title{Gumbel Random Number Generator}
\arguments{
\item{distrib}{A \code{GumbelDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers as \eqn{\mu - \sigma \log E} with \eqn{E} standard
exponential, which is inverse transform written without a logarithm of a
uniform.
}
\seealso{
\code{\link[=gumbel_distrib]{gumbel_distrib()}}
}
