% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distrib.R
\name{distrib_rng.LogisticDistrib}
\alias{distrib_rng.LogisticDistrib}
\title{Logistic Random Number Generator}
\arguments{
\item{distrib}{A \code{LogisticDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Logistic distribution.
}
\seealso{
\code{\link{logistic_distrib}}
}
