% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal1_distrib.R
\name{distrib_rng.Lognormal1Distrib}
\alias{distrib_rng.Lognormal1Distrib}
\title{Lognormal Random Number Generator}
\arguments{
\item{distrib}{A \code{Lognormal1Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Lognormal distribution.
}
\seealso{
\code{\link[=lognormal1_distrib]{lognormal1_distrib()}}
}
