% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distrib.R
\name{distrib_rng.LognormalDistrib}
\alias{distrib_rng.LognormalDistrib}
\title{Lognormal Random Number Generator}
\arguments{
\item{distrib}{A \code{LognormalDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Lognormal distribution.
}
\seealso{
\code{\link{lognormal_distrib}}
}
