% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin1_distrib.R
\name{distrib_rng.NegBin1Distrib}
\alias{distrib_rng.NegBin1Distrib}
\title{NB1 Random Generation}
\arguments{
\item{distrib}{A \code{NegBin1Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing \code{mu} and \code{theta}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
\code{\link[stats]{rnbinom}} at the same size and probability.
}
\seealso{
\code{\link{negbin1_distrib}}
}
