% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{distrib_rng.ReparamContinuousDistrib}
\alias{distrib_rng.ReparamContinuousDistrib}
\alias{reparam_rng}
\title{Random Generation from a Reparametrized Distribution}
\usage{
reparam_rng(distrib, n, theta)
}
\arguments{
\item{distrib}{A reparametrized distribution.}

\item{n}{The number of draws.}

\item{theta}{A named list of the new parameters.}
}
\value{
A numeric vector.
}
\description{
The parent's generator at the mapped parameters.
}
\seealso{
\code{\link{reparametrize}}
}
\keyword{internal}
