% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{distrib_rng.TransformedDistrib}
\alias{distrib_rng.TransformedDistrib}
\title{Transformed Random Number Generator}
\arguments{
\item{distrib}{A \code{TransformedDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list of the parent's parameters.}
}
\value{
A numeric vector of random draws.
}
\description{
Draws from the parent distribution and applies \eqn{g}.
}
\seealso{
\code{\link[=transformation]{transformation()}}
}
