% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution-functions}
\alias{distribution-functions}
\alias{pdf}
\alias{cdf}
\alias{qf}
\alias{rng}
\alias{likelihood}
\title{Distribution functions}
\usage{
pdf(distribution, x, log = FALSE)

cdf(distribution, q, lower.tail = TRUE, log.p = FALSE)

qf(distribution, p, lower.tail = TRUE, log.p = FALSE)

rng(distribution, n)

likelihood(distribution, x, log = TRUE, factor = 1)
}
\arguments{
\item{distribution}{an object of \code{S7} class \code{Distribution}.}

\item{x}{vector of quantiles.}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if \code{TRUE}, probabilities to the left of the quantile are given, otherwise to the right.}

\item{log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{factor}{numeric; factor by which to weight the total likelihood.}
}
\description{
Methods for distribution objects, providing the density (mass), cumulative probability, quantiles, and random number
generation.
}
