% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\alias{Distribution}
\alias{DistributionDiscrete}
\alias{DistributionContinuous}
\title{Distribution classes}
\usage{
Distribution(name = character(0), support = Real())

DistributionDiscrete(name = character(0), support = Int())

DistributionContinuous(name = character(0), support = Real())
}
\description{
These classes are abstract and provide the base classes for all individual distributions in the package.
These classes are not intended for direct use, unless you want to implement a custom distribution class.
}
