% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions7-package.R
\docType{package}
\name{distributions7-package}
\alias{distributions7}
\alias{distributions7-package}
\title{distributions7: S7 Framework for Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a consistent S7 object-oriented infrastructure for probability distributions. Each 'distrib' object encapsulates standard statistical functions alongside analytical derivatives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statmodels7.github.io/distributions7/}
  \item \url{https://github.com/statmodels7/distributions7}
  \item Report bugs at \url{https://github.com/statmodels7/distributions7/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Tinervia \email{giovannitinervia9@gmail.com}

}
\keyword{internal}
