% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-statistics.R
\name{fit-statistics}
\alias{fit-statistics}
\alias{information_criteria}
\alias{gof_test}
\alias{log_lik}
\alias{aic}
\alias{bic}
\alias{ks_test}
\alias{cvm_test}
\alias{ad_test}
\title{Fit statistics}
\usage{
information_criteria(distribution, data)

gof_test(
  distribution,
  data,
  estimated = FALSE,
  bootstrap = Bootstrap(samples = 0L)
)

log_lik(distribution, data)

aic(distribution, data)

bic(distribution, data)

ks_test(distribution, data)

cvm_test(distribution, data, estimated = FALSE)

ad_test(distribution, data, estimated = FALSE)
}
\arguments{
\item{distribution}{Object of class \code{\link[=Distribution]{Distribution()}}.}

\item{data}{Numeric containing the data.}

\item{estimated}{Was the distribution fitted to data on which we compute the goodness of fit tests? Passed to \link[goftest:ad.test]{goftest::ad.test} and \link[goftest:cvm.test]{goftest::cvm.test}.}

\item{bootstrap}{Object of class \code{\link[=Bootstrap]{Bootstrap()}}.}
}
\description{
Fit statistics of a distribution.
}
\details{
Calling \code{\link[=information_criteria]{information_criteria()}} and \code{\link[=gof_test]{gof_test()}} is recommended.

\code{\link[=information_criteria]{information_criteria()}} is valid only when the distribution has been already fitted to the data with \code{\link[=Mle]{Mle()}} (see \link{parameter-estimation}).

\code{\link[=gof_test]{gof_test()}} will work if the distribution has not been fitted to the data (\code{estimated=FALSE}), in which case it will compute
the Kolmogorov-Smirnov test and the simple hypothesis test Anderson-Darling and Cramer-von Mises.
When \code{estimated=TRUE}, the method will compute composite hypothesis test Anderson-Darling and Cramer-von Mises.
Alternatively, one can specify a \code{\link[=Bootstrap]{Bootstrap()}} method, which will perform a parametric bootstrap. Note that when \code{estimated=TRUE},
the \code{\link[=Estimator]{Estimator()}} that was used for fitting the distribution should be also passed in the \code{\link[=Bootstrap]{Bootstrap()}} option.
}
