% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{generate_random_theta.distrib}
\alias{generate_random_theta.distrib}
\title{Generate Random Parameters for \code{distrib} Objects}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A named list of randomly generated parameters.
}
\description{
Generates a named list of sensible random parameters based on the mathematical domain (\code{params_bounds}) of the given distribution.
}
\keyword{internal}
