% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexpected_hessian.R
\name{hess_pair_name}
\alias{hess_pair_name}
\title{The Hessian Name of a Pair of Parameters}
\usage{
hess_pair_name(params, a, b)
}
\arguments{
\item{params}{A character vector of parameter names.}

\item{a, b}{Indices into \code{params}.}
}
\value{
A single string.
}
\description{
\verb{params[a]_params[b]} in whichever order \code{\link[=hess_names]{hess_names()}} lists it.
}
\keyword{internal}
