% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-inverse_gamma.R
\name{inverse_gamma}
\alias{inverse_gamma}
\alias{InverseGammaScale}
\alias{InverseGammaRate}
\alias{InverseGammaMean}
\title{Inverse gamma distribution}
\usage{
inverse_gamma(alpha, theta, lambda, mu)

InverseGammaScale(alpha, theta)

InverseGammaRate(alpha, lambda)

InverseGammaMean(alpha, mu)
}
\arguments{
\item{alpha}{shape parameter.}

\item{theta}{scale parameter.}

\item{lambda}{rate parameter.}

\item{mu}{mean parameter.}
}
\description{
Create an inverse gamma distribution object.
}
\note{
Parameter meanings are on the inverse scale, e.g., \code{mu} is the mean of 1/x.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
