% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distInverseGamma.R
\name{InverseGamma}
\alias{InverseGamma}
\title{Inverse gamma distribution}
\usage{
InverseGamma(alpha, theta, lambda, mu)
}
\arguments{
\item{alpha}{shape parameter.}

\item{theta}{scale parameter.}

\item{lambda}{rate parameter.}

\item{mu}{mean parameter.}
}
\description{
Create an inverse gamma distribution object.
}
\note{
Parameter meanings are on the inverse scale, e.g., \code{mu} is the mean of 1/x.
}
\seealso{
Other distributions:
\code{\link[=Amoroso]{Amoroso()}},
\code{\link[=Beta]{Beta()}},
\code{\link[=BetaPrime]{BetaPrime()}},
\code{\link[=Binomial]{Binomial()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquared]{ChiSquared()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gompertz]{Gompertz()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=Laplace]{Laplace()}},
\code{\link[=LogLogistic]{LogLogistic()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=NegativeBinomial]{NegativeBinomial()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=Pareto]{Pareto()}},
\code{\link[=Shifted]{Shifted()}},
\code{\link[=SkewedGeneralizedT]{SkewedGeneralizedT()}},
\code{\link[=StretchedBeta]{StretchedBeta()}},
\code{\link[=SymmetricGeneralizedNormal]{SymmetricGeneralizedNormal()}},
\code{\link[=Triangular]{Triangular()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Wald]{Wald()}},
\code{\link[=Weibull]{Weibull()}},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
