% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{is_reparam}
\alias{is_reparam}
\title{Is This a Reparametrized Distribution?}
\usage{
is_reparam(distrib)
}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}
}
\value{
A single logical.
}
\description{
\code{TRUE} for either of the two wrapper classes.
}
\seealso{
\code{\link[=reparametrize]{reparametrize()}}
}
\keyword{internal}
