% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated.R
\name{is_zero_wrapper}
\alias{is_zero_wrapper}
\title{Does This Distribution Already Model a Probability of Zero?}
\usage{
is_zero_wrapper(distrib)
}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}
}
\value{
A single logical.
}
\description{
\code{TRUE} for a distribution produced by \code{\link{zero_inflated}} or
\code{\link{zero_adjusted}}, in either of its two forms.
}
\seealso{
\code{\link{check_not_stacked}}
}
\keyword{internal}
