% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kink_decomposition.R
\name{kink_decomposition.distrib}
\alias{kink_decomposition.distrib}
\title{The Base Method: No Kink}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}
}
\value{
\code{NULL}.
}
\description{
Returns \code{NULL}: the family is smooth in every parameter. Every family
inherits this except the three that declare an absolute value of the
residual in their log-density.
}
\examples{
kink_decomposition(gaussian1_distrib())

}
\seealso{
\code{\link[=kink_decomposition]{kink_decomposition()}} for the generic.
}
\keyword{internal}
