% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.Beta2Distrib}
\alias{kurtosis.Beta2Distrib}
\title{Kurtosis of the Beta in Its Shapes}
\arguments{
\item{x}{A \code{Beta2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form; excess.
}
\keyword{internal}
