% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.ChisqDistrib}
\alias{kurtosis.ChisqDistrib}
\title{Kurtosis of the Chi-Squared Distribution}
\arguments{
\item{x}{A \code{ChisqDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: excess \eqn{12/\mu}.
}
\keyword{internal}
