% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.distrib}
\alias{kurtosis.distrib}
\title{Kurtosis of a Distribution}
\arguments{
\item{x}{A \code{distrib} object.}

\item{theta}{A named list of parameters.}

\item{...}{Passed to \code{\link[=moment]{moment()}}.}
}
\value{
A numeric vector.
}
\description{
Computes the excess kurtosis, the standardized fourth central moment minus three through \code{\link[=moment]{moment()}}.
}
\keyword{internal}
