% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.Gamma2Distrib}
\alias{kurtosis.Gamma2Distrib}
\title{Kurtosis of the Gamma Distribution}
\arguments{
\item{x}{A \code{Gamma2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: excess \eqn{6/a = 6\sigma^2/\mu^2}.
}
\keyword{internal}
