% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.GeometricDistrib}
\alias{kurtosis.GeometricDistrib}
\title{Kurtosis of the Geometric Distribution}
\arguments{
\item{x}{A \code{GeometricDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: excess \eqn{6 + 1/\{\mu(1+\mu)\}}.
}
\keyword{internal}
