% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.InvGauss1Distrib}
\alias{kurtosis.InvGauss1Distrib}
\title{Kurtosis of the Inverse Gaussian Distribution}
\arguments{
\item{x}{An \code{InvGauss1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: excess \eqn{15\phi\mu}.
}
\keyword{internal}
