% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.SkewNormalDistrib}
\alias{kurtosis.SkewNormalDistrib}
\title{Kurtosis of the Skew Normal Distribution}
\arguments{
\item{x}{A \code{\link{SkewNormalDistrib}}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form, excess:
\eqn{2(\pi - 3)(b\delta)^4/(1 - b^2\delta^2)^2}. It is non-negative and
bounded above by about \eqn{0.8692}.
}
\keyword{internal}
