% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.WeibullDistrib}
\alias{kurtosis.WeibullDistrib}
\title{Kurtosis of the Weibull Distribution}
\arguments{
\item{x}{A \code{\link{WeibullDistrib}}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form, excess:
\eqn{(g_4 - 4g_1g_3 + 6g_1^2g_2 - 3g_1^4)/(g_2 - g_1^2)^2 - 3}, free of
the scale.
}
\keyword{internal}
