% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log_transform}
\alias{log_transform}
\title{Logarithmic Transformation}
\usage{
log_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \log(X)}: maps \eqn{(0, \infty)} to the real line.
Inverse \eqn{X = e^Y}, Jacobian \eqn{|J| = e^Y}.
}
