% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.Beta1Distrib}
\alias{mean.Beta1Distrib}
\title{Mean of the Beta Distribution}
\arguments{
\item{x}{A \code{Beta1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu}, a parameter of this parametrization.
}
\keyword{internal}
