% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.Gamma1Distrib}
\alias{mean.Gamma1Distrib}
\title{Mean of the Gamma in Mean and Dispersion}
\arguments{
\item{x}{A \code{Gamma1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu}, the parameter itself.
}
\keyword{internal}
