% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.Gaussian2Distrib}
\alias{mean.Gaussian2Distrib}
\title{Mean of the Gaussian in Mean and Variance}
\arguments{
\item{x}{A \code{Gaussian2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu}.
}
\keyword{internal}
