% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distrib.R
\name{mv_location.MultinomialDistrib}
\alias{mv_location.MultinomialDistrib}
\title{Multinomial Mean Vector}
\arguments{
\item{distrib}{A \code{MultinomialDistrib} object.}

\item{theta}{A named list of parameters.}
}
\value{
A numeric vector summing to the size.
}
\description{
\eqn{n p}, the size times the probability vector.
}
\seealso{
\code{\link[=multinomial_distrib]{multinomial_distrib()}}
}
