% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_distrib.R
\name{mv_sigma.DirichletDistrib}
\alias{mv_sigma.DirichletDistrib}
\title{Dirichlet Covariance Matrix}
\arguments{
\item{distrib}{A \code{DirichletDistrib} object.}

\item{theta}{A named list of parameters.}
}
\value{
A symmetric matrix.
}
\description{
\deqn{\operatorname{Cov}(Y_i, Y_j)
      = \dfrac{\delta_{ij}\mu_i - \mu_i\mu_j}{\phi + 1}}
singular by construction, the coordinates summing to a constant.
}
\seealso{
\code{\link{dirichlet_distrib}}
}
