% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distrib.R
\name{mv_sigma.MultinomialDistrib}
\alias{mv_sigma.MultinomialDistrib}
\title{Multinomial Covariance Matrix}
\arguments{
\item{distrib}{A \code{MultinomialDistrib} object.}

\item{theta}{A named list of parameters.}
}
\value{
A symmetric matrix.
}
\description{
\deqn{\operatorname{Cov}(Y_i, Y_j) = n(\delta_{ij}p_i - p_i p_j)}
singular by construction, the coordinates summing to the size.
}
\seealso{
\code{\link[=multinomial_distrib]{multinomial_distrib()}}
}
