% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-negative_binomial.R
\name{negative_binomial}
\alias{negative_binomial}
\alias{NegativeBinomialMean}
\alias{NegativeBinomialProb}
\title{Negative binomial distribution}
\usage{
negative_binomial(mu, phi, p, k)

NegativeBinomialMean(mu, phi)

NegativeBinomialProb(p, k)
}
\arguments{
\item{mu}{mean parameter.}

\item{phi}{dispersion parameter.}

\item{p}{probability of success parameter.}

\item{k}{number of successes parameter.}
}
\description{
Create a negative binomial distribution object.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
