% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{parameter-estimators}
\alias{parameter-estimators}
\alias{Estimator}
\alias{Mme}
\alias{Mle}
\alias{BiasCorrected}
\title{Estimators of distribution parameters.}
\usage{
Estimator(silent = FALSE)

Mme(silent = FALSE)

Mle(
  optim = FALSE,
  constrained = FALSE,
  method = if (constrained) "L-BFGS-B" else "BFGS",
  start = parameter_start,
  control = list(),
  silent = FALSE
)

BiasCorrected(silent = FALSE)
}
\arguments{
\item{silent}{logical; Should we suppress info messages during fitting.}

\item{optim}{logical; Should we force generic numerical optimization, even if there is a specific method to find the estimates for the particular distribution?}

\item{constrained}{logical; Should numerical optimization be done on the constrained or on the unconstrained parameter space?}

\item{method}{character; passed to \code{\link[stats:optim]{stats::optim()}}.}

\item{start}{method/function, list, or object of \code{S7} class \code{Estimator}. See details.}

\item{control}{list; passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
These classes encapsulate methods for parameter estimation.
}
\details{
These classes are used for dispatching methods such as \code{\link[=fit]{fit()}}, \code{\link[=parameter_estimates]{parameter_estimates()}}, or as a subroutine for \code{\link[=InferenceMethod]{InferenceMethod()}}.

\code{\link[=Mme]{Mme()}} specifies method of moments estimator, \code{\link[=Mle]{Mle()}}
specifies maximum likelihood estimator, and \code{\link[=BiasCorrected]{BiasCorrected()}} specifies some version of bias corrected estimators
(this is highly dependent on the type of distribution).

\code{\link[=Mle]{Mle()}} is the default method that is implemented for any distribution. If there are no closed-form solutions for the MLE,
then numerical optimization of the log likelihood is used.

\code{start} is used for initialization of the distribution parameters for numerical optimization.
By default, the routine calls \code{\link[=parameter_start]{parameter_start()}}, which for selected distributions attempts to use some heuristics to find
reasonable starting values, otherwise call \code{parameteter_estimates(distribution, Mme(), data)} to start the parameters
with method of moments estimation. In the absence of an implementation of either of the methods for a particular distribution,
the starting parameter values default to the values set by the distribution.
One can pass an arbitrary method/function with arguments \code{distribution}, \code{data}, to implement a custom parameter value initialization.
Otherwise, one can also pass a named list to overwrite the parameter values before starting the optimization.
}
