% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-methods.R
\name{parameter-inference}
\alias{parameter-inference}
\alias{InferenceMethod}
\alias{DefaultMethod}
\alias{NormalTheory}
\alias{ProfileLikelihood}
\alias{Bootstrap}
\title{Parameter inference methods}
\usage{
InferenceMethod(estimator = Mle(), ci_level = 0.95)

DefaultMethod(estimator = Mle(), ci_level = 0.95, args = list())

NormalTheory(
  estimator = Mle(),
  ci_level = 0.95,
  constrained = FALSE,
  control = list()
)

ProfileLikelihood(estimator=Mle(), ci_level=0.95)

Bootstrap(estimator=Mle(), ci_level=0.95, samples=1000L, callback=function() {})
}
\arguments{
\item{estimator}{Object of class \code{\link[=Estimator]{Estimator()}}.}

\item{ci_level}{numeric; confidence level.}

\item{args}{list; optional arguments for the method.}

\item{constrained}{logical; should the CIs be computed on the unconstrained space (and then transformed),
or should be computed directly on the constrained space.}

\item{control}{list; passed to \link[stats:optimHess]{stats::optimHess}.}

\item{samples}{integer; How many samples to take.}

\item{callback}{function; Optional function to exectute on every iteration.}
}
\description{
These classes encapsulate methods for parameter inference. This is yet experimental; methods for \code{\link[=Bootstrap]{Bootstrap()}} and \code{\link[=ProfileLikelihood]{ProfileLikelihood()}} are not implemented at all.
\code{\link[=DefaultMethod]{DefaultMethod()}} will currently default to \code{\link[=NormalTheory]{NormalTheory()}}, \emph{which is not a good default for some distributions!}
}
\seealso{
\code{\link[=parameter_inference]{parameter_inference()}}
}
