% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{parameter-properties}
\alias{parameter-properties}
\alias{parameters}
\alias{parameter_properties}
\alias{parameter_values}
\alias{parameter_values<-}
\alias{parameter_uvalues}
\alias{parameter_uvalues<-}
\alias{recreate_parameters}
\title{Parameter properties}
\usage{
parameters(distribution, which = c("all", "free", "fixed"))

parameter_properties(
  distribution,
  property = c("key", "name", "label", "value", "uvalue", "derivative", "support",
    "fixed"),
  which = "all"
)

parameter_values(distribution, which = "all")

parameter_values(distribution) <- value

parameter_uvalues(distribution, which = "all")

parameter_uvalues(distribution) <- value

recreate_parameters(distribution)
}
\arguments{
\item{distribution}{Object of class \code{\link[=Distribution]{Distribution()}}.}

\item{which}{character; should all, free, or fixed parameters be extracted?}

\item{property}{character; which parameter property to extract.}

\item{value}{list of parameter values (constrained or uncosntrained) to which to set the parameters.}
}
\description{
Get or set parameter properties. These methods are mostly intended for internal use, but can make some tasks easier elsewhere.
}
