% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\alias{Parameter}
\alias{fixed}
\alias{is.parameter}
\title{Parameter class}
\usage{
Parameter(key = "", name = key, label = name, value, support, fixed)

fixed(value)

is.parameter(x)
}
\arguments{
\item{key}{character; Parameter key.}

\item{name}{character; Optional name of the parameter.}

\item{label}{character; Optional label of the parameter. Typically a LaTeX expression (for convenience to render in software like in JASP).}

\item{value}{character; Value of the parameter.}

\item{support}{Object of class \code{\link[=Support]{Support()}}.}

\item{fixed}{logical; Is the parameter fixed? If true, the parameter will not be estimated with parameter estimation methods.}
}
\description{
Parameter classes encapsulating important information about distribution parameters.
This class is used internally and is not suppored to by used directly, unless one wants to implement a custom distribution.
}
\details{
Fixed parameter can be either specified by setting the \code{fixed} property to \code{TRUE}, but one can also fix the parameter by
using the \code{fixed()} function on the parameter value. For example \code{normal(mu=0, sigma=fixed(1))} specifies a normal distribution
with the standard deviation parameter fixed to 1.
}
