% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_normal_higher.R
\name{phi_hermite}
\alias{phi_hermite}
\title{Derivatives of the Standard Normal Distribution Function}
\usage{
phi_hermite(x, m)
}
\arguments{
\item{x}{A numeric vector.}

\item{m}{The order, 1 to 4.}
}
\value{
A numeric vector.
}
\description{
Returns the factor \eqn{h_m} in
\eqn{\Phi^{(m)}(x) = h_m(x)\varphi(x)} for \eqn{m = 1, \ldots, 4}, the
Hermite polynomials.
}
\keyword{internal}
