% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.continuous_distrib}
\alias{plot.continuous_distrib}
\title{Plot Method for Continuous Distributions}
\arguments{
\item{x}{An object of class \code{"continuous_distrib"}.}

\item{theta}{A named list or vector of parameters matching \code{x@params}.}

\item{xlim}{Optional numeric vector of length 2 indicating the x-axis range.}

\item{...}{Additional arguments passed to the base \code{\link{plot}} function.}
}
\description{
Visualizes the Probability Density Function (PDF) of a continuous distribution object.
}
\keyword{internal}
