% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distrib.R
\name{print.distrib_fit}
\alias{print.distrib_fit}
\title{Print Method for Maximum-Likelihood Fits}
\arguments{
\item{x}{A \code{\link[=distrib_fit]{distrib_fit()}} object.}

\item{digits}{Number of significant digits. Defaults to 4.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for Maximum-Likelihood Fits
}
