% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distSkew.R
\name{skew-distributions}
\alias{skew-distributions}
\alias{SkewNormal}
\alias{SkewCauchy}
\alias{SkewT}
\title{Skew distribution family}
\usage{
SkewNormal(xi, omega, alpha)

SkewCauchy(xi, omega, alpha)

SkewT(xi, omega, alpha, nu)
}
\arguments{
\item{xi}{location parameter.}

\item{omega}{scale parameter.}

\item{alpha}{slant (skew) parameter.}

\item{nu}{degrees of freedom parameter.}
}
\description{
Create an object of the Skew distribution family.
}
\seealso{
Other distributions:
\code{\link[=Amoroso]{Amoroso()}},
\code{\link[=Beta]{Beta()}},
\code{\link[=BetaPrime]{BetaPrime()}},
\code{\link[=Binomial]{Binomial()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquared]{ChiSquared()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gompertz]{Gompertz()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=InverseGamma]{InverseGamma()}},
\code{\link[=Laplace]{Laplace()}},
\code{\link[=LogLogistic]{LogLogistic()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=NegativeBinomial]{NegativeBinomial()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=Pareto]{Pareto()}},
\code{\link[=Shash]{Shash()}},
\code{\link[=Shifted]{Shifted()}},
\code{\link[=SkewedGeneralizedT]{SkewedGeneralizedT()}},
\code{\link[=StretchedBeta]{StretchedBeta()}},
\code{\link[=SymmetricGeneralizedNormal]{SymmetricGeneralizedNormal()}},
\code{\link[=Triangular]{Triangular()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Wald]{Wald()}},
\code{\link[=Weibull]{Weibull()}},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{t-distribution}}
}
\concept{distributions}
