% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skew.R
\name{skew}
\alias{skew}
\alias{skew_normal}
\alias{skew_cauchy}
\alias{skew_t}
\alias{SkewNormal}
\alias{SkewCauchy}
\alias{SkewT}
\title{Skew distribution family}
\usage{
skew_normal(xi, omega, alpha)

skew_cauchy(xi, omega, alpha)

skew_t(xi, omega, alpha, nu)

SkewNormal(xi, omega, alpha)

SkewCauchy(xi, omega, alpha)

SkewT(xi, omega, alpha, nu)
}
\arguments{
\item{xi}{location parameter.}

\item{omega}{scale parameter.}

\item{alpha}{slant (skew) parameter.}

\item{nu}{degrees of freedom parameter.}
}
\description{
Create an object of Skew family.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
