% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewed_generalized_t.R
\name{skewed_generalized_t}
\alias{skewed_generalized_t}
\alias{SkewedGeneralizedT}
\title{Skewed generalized t-distribution}
\usage{
skewed_generalized_t(mu, sigma, lambda, p, q)

SkewedGeneralizedT(mu, sigma, lambda, p, q)
}
\arguments{
\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{lambda}{skew parameter.}

\item{p}{kurtosis parameter.}

\item{q}{kurtosis parameter.}
}
\description{
Create a skewed generalized t-distribution.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
