% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.Beta2Distrib}
\alias{skewness.Beta2Distrib}
\title{Skewness of the Beta in Its Shapes}
\arguments{
\item{x}{A \code{Beta2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form:
\eqn{2(\beta-\alpha)\sqrt{\alpha+\beta+1}/
       \{(\alpha+\beta+2)\sqrt{\alpha\beta}\}}.
}
\keyword{internal}
