% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.ExponentialDistrib}
\alias{skewness.ExponentialDistrib}
\title{Skewness of the Exponential Distribution}
\arguments{
\item{x}{An \code{ExponentialDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: 2, free of the parameter.
}
\keyword{internal}
