% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.Gamma1Distrib}
\alias{skewness.Gamma1Distrib}
\title{Skewness of the Gamma in Mean and Dispersion}
\arguments{
\item{x}{A \code{Gamma1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{2\sqrt{\phi}}, free of the mean.
}
\keyword{internal}
