% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.Gamma2Distrib}
\alias{skewness.Gamma2Distrib}
\title{Skewness of the Gamma Distribution}
\arguments{
\item{x}{A \code{Gamma2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{2/\sqrt{a}} with shape
\eqn{a = \mu^2/\sigma^2}, hence \eqn{2\sigma/\mu}.
}
\keyword{internal}
