% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.Gaussian3Distrib}
\alias{skewness.Gaussian3Distrib}
\title{Skewness of the Gaussian in Mean and Precision}
\arguments{
\item{x}{A \code{Gaussian3Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: zero.
}
\keyword{internal}
