% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.InvGauss1Distrib}
\alias{skewness.InvGauss1Distrib}
\title{Skewness of the Inverse Gaussian Distribution}
\arguments{
\item{x}{An \code{InvGauss1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{3\sqrt{\phi\mu}}.
}
\keyword{internal}
