% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.numeric}
\alias{skewness.numeric}
\title{Sample Skewness}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Unused.}
}
\value{
A single number.
}
\description{
The sample skewness of a numeric vector, computed from the sample's central moments.
}
\keyword{internal}
