% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.PoissonDistrib}
\alias{skewness.PoissonDistrib}
\title{Skewness of the Poisson Distribution}
\arguments{
\item{x}{A \code{PoissonDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{1/\sqrt{\mu}}.
}
\keyword{internal}
