% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.WeibullDistrib}
\alias{skewness.WeibullDistrib}
\title{Skewness of the Weibull Distribution}
\arguments{
\item{x}{A \code{\link{WeibullDistrib}}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form:
\eqn{(g_3 - 3g_1g_2 + 2g_1^3)/(g_2 - g_1^2)^{3/2}}, free of the scale.
}
\keyword{internal}
