% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{std_dev.numeric}
\alias{std_dev.numeric}
\title{Sample Standard Deviation}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Remove missing values first?}

\item{...}{Unused.}
}
\value{
A single number.
}
\description{
The sample standard deviation of a numeric vector, delegated to \code{\link[stats:sd]{stats::sd()}}.
}
\keyword{internal}
