% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-stretched_beta.R
\name{stretched_beta}
\alias{stretched_beta}
\alias{StretchedBeta}
\title{Stretched beta distribution}
\usage{
stretched_beta(alpha, beta, min = fixed(0), max = fixed(1))

StretchedBeta(alpha, beta, min, max)
}
\arguments{
\item{alpha}{shape parameter.}

\item{beta}{shape parameter.}

\item{min}{minimum parameter.}

\item{max}{maximum parameter.}
}
\description{
Create a stretched beta distribution object.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
