% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t2_distrib.R
\name{student_t2_scale}
\alias{student_t2_scale}
\title{The Scale of a Student t with a Given Standard Deviation}
\usage{
student_t2_scale(theta)
}
\arguments{
\item{theta}{A list with the location, the standard deviation and the
degrees of freedom, in that order.}
}
\value{
A numeric vector of scales.
}
\description{
Returns \eqn{s_0 = \sigma\sqrt{1 - 2/\nu}}, the scale of the
Student t whose standard deviation is \eqn{\sigma}.
}
\seealso{
\code{\link[=distrib_pdf.StudentT2Distrib]{distrib_pdf.StudentT2Distrib()}}
}
\keyword{internal}
