% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{support-methods}
\alias{support-methods}
\alias{support}
\alias{inside}
\alias{unconstrain}
\alias{constrain}
\alias{derivative}
\title{Methods for support classes.}
\usage{
support(object, ...)

inside(object, ...)

unconstrain(object, x, ...)

constrain(object, x, ...)

derivative(object, x, ...)
}
\arguments{
\item{object}{Object of class \code{\link[=Support]{Support()}}, \code{\link[=Parameter]{Parameter()}}, or \code{\link[=Distribution]{Distribution()}}.}

\item{x}{Vector or values.}
}
\value{
\code{\link[=support]{support()}} returns the \link[=support]{Support()} of the object (either the object itself or the \code{support} property). For \code{\link[=Distribution]{Distribution()}},
the returned support object will always be \code{support@numeric == TRUE}, evaluating \code{min} and \code{max} properties if they are
specified as expressions. This allows you to inpect support of a distribution for distributions whose support depends on its
parameter values.

\code{\link[=inside]{inside()}} returns a logical vector indicating whether \code{x} is inside of the support. Note: support limits are always
treated as inclusive; one has to treat boundary conditions manually.

\code{\link[=constrain]{constrain()}} returns a numeric vector transforming \code{x} from unconstrained space into a constrained space.
\code{\link[=unconstrain]{unconstrain()}} returns a numeric vector transforming \code{x} from a constrained space into an unconstrained space.
\code{\link[=derivative]{derivative()}} returns a derivative of the constrain transform at the unconstrained value \code{x}.
}
\description{
Methods for \code{\link[=Support]{Support()}} classes.
}
