% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{Support}
\alias{Support}
\alias{Real}
\alias{Int}
\alias{support}
\alias{inside}
\alias{unconstrain}
\alias{constrain}
\alias{derivative}
\title{Support classes}
\usage{
Support(min = -Inf, max = Inf)

Real(min = -Inf, max = Inf)

Int(min = -Inf, max = Inf)

support(object)

inside(object, x)

unconstrain(object, x)

constrain(object, x)

derivative(object, x)
}
\arguments{
\item{min}{Minimum of the support. Can be either a numeric, or an expression.}

\item{max}{Maximum of the support. Can be either a numeric, or an expression.}

\item{object}{Object of class \code{\link[=Support]{Support()}}, \code{\link[=Parameter]{Parameter()}}, or \code{\link[=Distribution]{Distribution()}}.}

\item{x}{Vector or values.}
}
\value{
\code{\link[=support]{support()}} returns the \link[=support]{Support()} of the object (either the object itself or the \code{support} property). For \code{\link[=Distribution]{Distribution()}},
the returned support object will always be \code{support@numeric == TRUE}, evaluating \code{min} and \code{max} properties if they are
specified as expressions. This allows you to inpect support of a distribution for distributions whose support depends on its
parameter values.

\code{\link[=inside]{inside()}} returns a logical vector indicating whether \code{x} is inside of the support. Note: support limits are always
treated as inclusive; one has to treat boundary conditions manually.

\code{\link[=constrain]{constrain()}} returns a numeric vector transforming \code{x} from unconstrained space into a constrained space.
\code{\link[=unconstrain]{unconstrain()}} returns a numeric vector transforming \code{x} from a constrained space into an unconstrained space.
\code{\link[=derivative]{derivative()}} returns a derivative of the constrain transform at the unconstrained value \code{x}.
}
\description{
Classes the encapsulate support of distributions or parameters.
}
