% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-symmetric_generalized_normal.R
\name{symmetric_generalized_normal}
\alias{symmetric_generalized_normal}
\alias{SymmetricGeneralizedNormal}
\title{Symmetric generalized normal distribution (Subbotin)}
\usage{
symmetric_generalized_normal(mu, alpha, beta)

SymmetricGeneralizedNormal(mu, alpha, beta)
}
\arguments{
\item{mu}{location parameter.}

\item{alpha}{scale parameter.}

\item{beta}{shape parameter.}
}
\description{
Create a symmetric generalized normal distribution object.
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{distributions}},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
