% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNoncentralStudentT.R
\name{t-distribution}
\alias{t-distribution}
\alias{StandardT}
\alias{StudentT}
\alias{NoncentralT}
\alias{NoncentralStudentT}
\title{T-distribution}
\usage{
StandardT(nu)

StudentT(nu, mu, sigma)

NoncentralT(nu, kappa)

NoncentralStudentT(nu, kappa, mu, sigma)
}
\arguments{
\item{nu}{degrees of freedom parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{kappa}{noncentrality parameter.}
}
\description{
Create a t-distribution object.
}
\seealso{
Other distributions:
\code{\link[=Amoroso]{Amoroso()}},
\code{\link[=Beta]{Beta()}},
\code{\link[=BetaPrime]{BetaPrime()}},
\code{\link[=Binomial]{Binomial()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquared]{ChiSquared()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gompertz]{Gompertz()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=InverseGamma]{InverseGamma()}},
\code{\link[=Laplace]{Laplace()}},
\code{\link[=LogLogistic]{LogLogistic()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=NegativeBinomial]{NegativeBinomial()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=Pareto]{Pareto()}},
\code{\link[=Shifted]{Shifted()}},
\code{\link[=SkewedGeneralizedT]{SkewedGeneralizedT()}},
\code{\link[=StretchedBeta]{StretchedBeta()}},
\code{\link[=SymmetricGeneralizedNormal]{SymmetricGeneralizedNormal()}},
\code{\link[=Triangular]{Triangular()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Wald]{Wald()}},
\code{\link[=Weibull]{Weibull()}},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}}
}
\concept{distributions}
