% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformer}
\alias{transformer}
\title{S7 Class for Variable Transformers}
\usage{
transformer(
  name = character(0),
  trans_fun = function() NULL,
  trans_inv = function() NULL,
  trans_abs_jac = function() NULL,
  trans_inv_hessian = function() NULL,
  grad_log_jac = function() NULL,
  hess_log_jac = function() NULL,
  bounds_fun = function() NULL,
  valid_support = function() NULL,
  decreasing = logical(0)
)
}
\arguments{
\item{name}{A character string identifying the transformation.}

\item{trans_fun}{The forward transformation \eqn{y = g(x)}.}

\item{trans_inv}{The inverse transformation \eqn{x = g^{-1}(y)}.}

\item{trans_abs_jac}{The absolute Jacobian of the inverse transformation
\eqn{|J(y)| = |dx/dy|}; must accept a \code{log} argument.}

\item{trans_inv_hessian}{The second derivative of the inverse transformation \eqn{d^2x/dy^2}.}

\item{grad_log_jac}{The first derivative of \eqn{\log|J(y)|} with respect to \eqn{y}.}

\item{hess_log_jac}{The second derivative of \eqn{\log|J(y)|} with respect to \eqn{y}.}

\item{bounds_fun}{Maps the original support bounds to the transformed ones.}

\item{valid_support}{Checks whether a support is compatible with the transformation.}

\item{decreasing}{Logical; \code{TRUE} for monotonically decreasing transformations.}
}
\description{
A \code{transformer} object defines the mathematical rules for transforming a random
variable \eqn{Y = g(X)}. It is used as input to \code{\link{transformation}}.
}
\seealso{
\code{\link{transformation}}
}
