% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.BernoulliDistrib}
\alias{variance.BernoulliDistrib}
\title{Variance of the Bernoulli Distribution}
\arguments{
\item{x}{A \code{BernoulliDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu(1-\mu)}.
}
\keyword{internal}
