% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.Beta2Distrib}
\alias{variance.Beta2Distrib}
\title{Variance of the Beta in Its Shapes}
\arguments{
\item{x}{A \code{Beta2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form:
\eqn{\alpha\beta/\{(\alpha+\beta)^2(\alpha+\beta+1)\}}.
}
\keyword{internal}
