% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.CauchyDistrib}
\alias{variance.CauchyDistrib}
\title{The Cauchy Distribution Has No Variance}
\arguments{
\item{x}{A \code{CauchyDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector of \code{NaN}.
}
\description{
\code{NaN}; see \code{\link[=mean.CauchyDistrib]{mean()}}.
}
\keyword{internal}
