% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.ChisqDistrib}
\alias{variance.ChisqDistrib}
\title{Variance of the Chi-Squared Distribution}
\arguments{
\item{x}{A \code{ChisqDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{2\mu}.
}
\keyword{internal}
