% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.Gaussian3Distrib}
\alias{variance.Gaussian3Distrib}
\title{Variance of the Gaussian in Mean and Precision}
\arguments{
\item{x}{A \code{Gaussian3Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{1/\tau}.
}
\keyword{internal}
