% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.GeometricDistrib}
\alias{variance.GeometricDistrib}
\title{Variance of the Geometric Distribution}
\arguments{
\item{x}{A \code{GeometricDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu(1+\mu)}.
}
\keyword{internal}
