% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.numeric}
\alias{variance.numeric}
\title{Sample Variance}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Remove missing values first?}

\item{...}{Unused.}
}
\value{
A single number.
}
\description{
The sample variance of a numeric vector, delegated to \code{\link[stats:cor]{stats::var()}}.
}
\keyword{internal}
