% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal2_distrib.R
\name{variance.SkewNormal2Distrib}
\alias{variance.SkewNormal2Distrib}
\title{Variance of the Skew Normal in the Centered Parametrization}
\arguments{
\item{x}{A \code{SkewNormal2Distrib} object.}

\item{theta}{A list with \code{mu}, \code{sigma} and \code{gamma1}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\sigma^2}, the square of the parameter.
}
\keyword{internal}
