% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{weights_ic}
\alias{weights_ic}
\title{Compute weights from information criteria}
\usage{
weights_ic(x)
}
\arguments{
\item{x}{Numeric vector of AIC/BIC values.}
}
\value{
A vector of weights.
}
\description{
Compute weights from \code{\link[=aic]{aic()}} or \code{\link[=bic]{bic()}} values.
}
