% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{asinh_transform}
\alias{asinh_transform}
\title{Inverse Hyperbolic Sine Transformation}
\usage{
asinh_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \text{asinh}(X)}, a log-like transformation that
handles zero and negative values. Inverse \eqn{X = \sinh(Y)}, Jacobian \eqn{|J| = \cosh(Y)}.
}
