% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{asinh_transform}
\alias{asinh_transform}
\title{Inverse Hyperbolic Sine Transformation}
\usage{
asinh_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \text{asinh}(X)}, a log-like transformation that
handles zero and negative values. Inverse \eqn{X = \sinh(Y)}, Jacobian \eqn{|J| = \cosh(Y)}.
}
\examples{
d <- transformation(gaussian1_distrib(), asinh_transform())
distrib_pdf(d, 1, list(mu = 0, sigma = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{bc_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{logit_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
