% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distBinomial.R
\name{Binomial}
\alias{Binomial}
\title{Binomial distribution}
\usage{
Binomial(prob, size)
}
\arguments{
\item{prob}{Probability of success parameter.}

\item{size}{Sample size parameter.}
}
\description{
Create a binomial distribution object.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
