% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distChiSquared.R
\name{ChiSquared}
\alias{ChiSquared}
\alias{NoncentralChiSquared}
\title{Chi-squared distribution}
\usage{
ChiSquared(nu)

NoncentralChiSquared(nu, kappa)
}
\arguments{
\item{nu}{degrees of freedom paramter.}

\item{kappa}{noncentrality parameter.}
}
\description{
Create a chi-squared distribution object.
}
\seealso{
Other distributions:
\code{\link[=Amoroso]{Amoroso()}},
\code{\link[=Beta]{Beta()}},
\code{\link[=BetaPrime]{BetaPrime()}},
\code{\link[=Binomial]{Binomial()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gompertz]{Gompertz()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=InverseGamma]{InverseGamma()}},
\code{\link[=Laplace]{Laplace()}},
\code{\link[=LogLogistic]{LogLogistic()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=NegativeBinomial]{NegativeBinomial()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=Pareto]{Pareto()}},
\code{\link[=Shash]{Shash()}},
\code{\link[=Shifted]{Shifted()}},
\code{\link[=SkewedGeneralizedT]{SkewedGeneralizedT()}},
\code{\link[=StretchedBeta]{StretchedBeta()}},
\code{\link[=SymmetricGeneralizedNormal]{SymmetricGeneralizedNormal()}},
\code{\link[=Triangular]{Triangular()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Wald]{Wald()}},
\code{\link[=Weibull]{Weibull()}},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
