% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distrib.R
\name{coef.distrib_fit}
\alias{coef.distrib_fit}
\title{Extract Estimates from a Maximum-Likelihood Fit}
\arguments{
\item{object}{A \code{\link{distrib_fit}} object.}

\item{scale}{Either \code{"parameter"} (default) or \code{"link"}.}

\item{...}{Unused.}
}
\value{
A named numeric vector of estimates.
}
\description{
Extract Estimates from a Maximum-Likelihood Fit
}
