% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{discrete_distrib}
\alias{discrete_distrib}
\title{S7 Class for Discrete Distributions}
\usage{
discrete_distrib(
  distrib_name = character(0),
  dimension = character(0),
  bounds = integer(0),
  params = character(0),
  params_interpretation = character(0),
  n_params = integer(0),
  params_bounds = list(),
  link_params = list()
)
}
\arguments{
\item{distrib_name}{A single character string specifying the name of the distribution (e.g., \code{"student t"}).}

\item{dimension}{A character string indicating the dimensionality (\code{"univariate"} or \code{"multivariate"}).}

\item{bounds}{A numeric vector of length 2 defining the overall support of the distribution \code{c(lower, upper)}.}

\item{params}{A character vector containing the names of the distribution parameters (e.g., \code{c("mu", "sigma")}).}

\item{params_interpretation}{A character vector (typically named) providing the statistical interpretation of each parameter (e.g., \code{c(mu = "location")}).}

\item{n_params}{A numeric value specifying the total number of parameters.}

\item{params_bounds}{A list of numeric vectors of length 2, specifying the valid mathematical domain for each individual parameter.}

\item{link_params}{A list of link function objects corresponding to each parameter, primarily used to map parameters to the unconstrained real line for optimization algorithms.}
}
\description{
A subclass of \code{distrib} specifically for discrete probability distributions.
}
