% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{distrib_atoms.ReparamContinuousDistrib}
\alias{distrib_atoms.ReparamContinuousDistrib}
\alias{reparam_atoms}
\title{Atoms of a Reparametrized Distribution}
\usage{
reparam_atoms(distrib, theta)
}
\arguments{
\item{distrib}{A reparametrized distribution.}

\item{theta}{A named list of the new parameters.}
}
\value{
A numeric vector of atom locations, possibly empty.
}
\description{
The parent's, unchanged: a reparametrization does not move mass.
}
\seealso{
\code{\link[=reparametrize]{reparametrize()}}
}
\keyword{internal}
