% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distrib.R
\name{distrib_cdf.BernoulliDistrib}
\alias{distrib_cdf.BernoulliDistrib}
\title{Bernoulli Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{BernoulliDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the cumulative distribution function for the Bernoulli distribution.
}
\seealso{
\code{\link{bernoulli_distrib}}
}
