% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distrib.R
\name{distrib_cdf.CauchyDistrib}
\alias{distrib_cdf.CauchyDistrib}
\title{Cauchy Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{CauchyDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the cumulative distribution function for the Cauchy distribution.
}
\seealso{
\code{\link{cauchy_distrib}}
}
