% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_distrib.R
\name{distrib_cdf.Gamma1Distrib}
\alias{distrib_cdf.Gamma1Distrib}
\title{Gamma Distribution Function in Mean and Dispersion}
\arguments{
\item{distrib}{A \code{Gamma1Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list with \code{mu} and \code{phi}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The gamma distribution function at the implied shape and rate.
}
\seealso{
\code{\link[=gamma1_distrib]{gamma1_distrib()}}
}
