% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distrib.R
\name{distrib_cdf.GammaDistrib}
\alias{distrib_cdf.GammaDistrib}
\title{Gamma Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{GammaDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the cumulative distribution function for the Gamma distribution.
}
\seealso{
\code{\link{gamma_distrib}}
}
