% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian3_distrib.R
\name{distrib_cdf.Gaussian3Distrib}
\alias{distrib_cdf.Gaussian3Distrib}
\title{Gaussian Distribution Function in Mean and Precision}
\arguments{
\item{distrib}{A \code{Gaussian3Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list with \code{mu} and \code{tau}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The normal distribution function at standard deviation
\eqn{1/\sqrt{\tau}}.
}
\seealso{
\code{\link{gaussian3_distrib}}
}
