% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distrib.R
\name{distrib_cdf.GeometricDistrib}
\alias{distrib_cdf.GeometricDistrib}
\title{Geometric Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{GeometricDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities are returned as logarithms.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
\deqn{F(q; \mu) = 1 - \left(\dfrac{\mu}{1+\mu}\right)^{\lfloor q \rfloor + 1}}
}
\seealso{
\code{\link{geometric_distrib}}
}
