% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss2_distrib.R
\name{distrib_cdf.InvGauss2Distrib}
\alias{distrib_cdf.InvGauss2Distrib}
\title{Inverse Gaussian Distribution Function in Mean and Shape}
\arguments{
\item{distrib}{An \code{InvGauss2Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list with \code{mu} and \code{lambda}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The inverse gaussian distribution function at dispersion
\eqn{1/\lambda}.
}
\seealso{
\code{\link[=invgauss2_distrib]{invgauss2_distrib()}}
}
