% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss_distrib.R
\name{distrib_cdf.InvGaussDistrib}
\alias{distrib_cdf.InvGaussDistrib}
\title{Inverse-Gaussian Cumulative Distribution Function}
\arguments{
\item{distrib}{An \code{InvGaussDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the cumulative distribution function for the Inverse-Gaussian distribution.
}
\seealso{
\code{\link{invgauss_distrib}}
}
