% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace2_distrib.R
\name{distrib_cdf.Laplace2Distrib}
\alias{distrib_cdf.Laplace2Distrib}
\title{Laplace Distribution Function in Location and Rate}
\arguments{
\item{distrib}{A \code{Laplace2Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu} and \code{lambda}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
\deqn{F(q; \mu, \lambda) = \begin{cases} \dfrac{1}{2}\exp\left(\lambda(q-\mu)\right) & q < \mu \\ 1 - \dfrac{1}{2}\exp\left(-\lambda(q-\mu)\right) & q \ge \mu \end{cases}}
}
\seealso{
\code{\link[=laplace2_distrib]{laplace2_distrib()}}
}
