% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distrib.R
\name{distrib_cdf.LogisticDistrib}
\alias{distrib_cdf.LogisticDistrib}
\title{Logistic Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{LogisticDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Computes the cumulative distribution function for the Logistic distribution:
\deqn{F(q; \mu, \sigma) = \dfrac{1}{1 + \exp\left(-\dfrac{q-\mu}{\sigma}\right)}}
}
\seealso{
\code{\link[=logistic_distrib]{logistic_distrib()}}
}
