% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin1_distrib.R
\name{distrib_cdf.NegBin1Distrib}
\alias{distrib_cdf.NegBin1Distrib}
\title{NB1 Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{NegBin1Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing \code{mu} and \code{theta}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities are returned as logarithms.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
\code{\link[stats]{pnbinom}} at size \eqn{\mu/\theta} and
probability \eqn{1/(1+\theta)}.
}
\seealso{
\code{\link{negbin1_distrib}}
}
