% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distrib_cdf}
\alias{distrib_cdf}
\title{Cumulative Distribution Function}
\usage{
distrib_cdf(distrib, ...)
}
\arguments{
\item{distrib}{A distribution object inheriting from the \code{distrib} class.}

\item{...}{Additional arguments passed to the specific method (e.g., \code{q}, \code{theta}, \code{lower.tail}).}
}
\description{
Evaluates the cumulative distribution function (CDF) for a given distribution.
}
