% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal2_distrib.R
\name{distrib_cdf.SkewNormal2Distrib}
\alias{distrib_cdf.SkewNormal2Distrib}
\title{Skew Normal Distribution Function in the Centered Parametrization}
\arguments{
\item{distrib}{A \code{SkewNormal2Distrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list with \code{mu}, \code{sigma} and \code{gamma1}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{...}{Passed on.}
}
\value{
A numeric vector.
}
\description{
The skew normal distribution function, through Owen's T.
}
\seealso{
\code{\link[=skewnormal2_distrib]{skewnormal2_distrib()}}
}
