% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t_distrib.R
\name{distrib_cdf.StudentTDistrib}
\alias{distrib_cdf.StudentTDistrib}
\title{Student's t Cumulative Distribution Function}
\arguments{
\item{distrib}{A \code{StudentTDistrib} object.}

\item{q}{A numeric vector of quantiles.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma}, and \code{nu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le q)}, otherwise \eqn{P(Y > q)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the cumulative distribution function for the Student's t distribution.
}
\seealso{
\code{\link{student_t_distrib}}
}
