% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_derivatives.R
\name{distrib_deriv4.TruncatedContinuousDistrib}
\alias{distrib_deriv4.TruncatedContinuousDistrib}
\title{Truncated Fourth Derivatives (Continuous)}
\arguments{
\item{distrib}{A \code{TruncatedContinuousDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list of the parent's parameters.}

\item{expected}{Logical; if \code{TRUE}, the expected derivatives.}
}
\value{
A named list of derivative components.
}
\description{
As \code{\link[=distrib_deriv3.TruncatedContinuousDistrib]{the third()}}, at fourth order.
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
