% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distrib.R
\name{distrib_expected_hessian.BernoulliDistrib}
\alias{distrib_expected_hessian.BernoulliDistrib}
\title{Bernoulli Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{BernoulliDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Bernoulli log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\dfrac{1}{\mu(1-\mu)}}
}
\seealso{
\code{\link[=bernoulli_distrib]{bernoulli_distrib()}}
}
