% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta1_distrib.R
\name{distrib_expected_hessian.Beta1Distrib}
\alias{distrib_expected_hessian.Beta1Distrib}
\title{Beta Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{Beta1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Beta log-density
with respect to the parameters \eqn{\mu} and \eqn{\phi}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\phi^2 \left[ \psi_1(\mu\phi) + \psi_1((1-\mu)\phi) \right]}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \phi^2}\right] = \psi_1(\phi) - \mu^2\psi_1(\mu\phi) - (1-\mu)^2\psi_1((1-\mu)\phi)}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu \partial \phi}\right] = -\phi \left[ \mu\psi_1(\mu\phi) - (1-\mu)\psi_1((1-\mu)\phi) \right]}
}
\seealso{
\code{\link{beta1_distrib}}
}
