% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distrib.R
\name{distrib_expected_hessian.BinomialDistrib}
\alias{distrib_expected_hessian.BinomialDistrib}
\title{Binomial Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{BinomialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Binomial log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\dfrac{n}{\mu(1-\mu)}}
}
\seealso{
\code{\link{binomial_distrib}}
}
