% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distrib.R
\name{distrib_expected_hessian.CauchyDistrib}
\alias{distrib_expected_hessian.CauchyDistrib}
\title{Cauchy Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{CauchyDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Cauchy log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = -\dfrac{1}{2\sigma^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \sigma^2}\right] = -\dfrac{1}{2\sigma^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma}\right] = 0}
}
\seealso{
\code{\link[=cauchy_distrib]{cauchy_distrib()}}
}
