% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2_distrib.R
\name{distrib_expected_hessian.Gamma2Distrib}
\alias{distrib_expected_hessian.Gamma2Distrib}
\title{Gamma Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{Gamma2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Gamma log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma^2}.

\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu^2}\right] = \dfrac{3\sigma^2 - 4\mu^2\psi_1\left(\dfrac{\mu^2}{\sigma^2}\right)}{(\sigma^2)^2}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial (\sigma^2)^2}\right] = -\dfrac{\mu^2\left(\mu^2\psi_1\left(\dfrac{\mu^2}{\sigma^2}\right) - \sigma^2\right)}{(\sigma^2)^4}}
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma^2}\right] = \dfrac{2\mu\left(\mu^2\psi_1\left(\dfrac{\mu^2}{\sigma^2}\right) - \sigma^2\right)}{(\sigma^2)^3}}
}
\seealso{
\code{\link{gamma2_distrib}}
}
