% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal1_distrib.R
\name{distrib_expected_hessian.Lognormal1Distrib}
\alias{distrib_expected_hessian.Lognormal1Distrib}
\title{Lognormal Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{Lognormal1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Computes the analytical expected Hessian of the Lognormal log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma^2}.
}
\seealso{
\code{\link[=lognormal1_distrib]{lognormal1_distrib()}}
}
