% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distrib_expected_hessian}
\alias{distrib_expected_hessian}
\title{Analytical Expected Hessian}
\usage{
distrib_expected_hessian(distrib, ...)
}
\arguments{
\item{distrib}{A distribution object inheriting from the \code{distrib} class.}

\item{...}{Additional arguments passed to the specific method (e.g., \code{y}, \code{theta}).}
}
\description{
Computes the analytical expected second derivatives of the log-likelihood with respect to the distribution's parameters.
}
