% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{distrib_expected_hessian.TruncatedDiscreteDistrib}
\alias{distrib_expected_hessian.TruncatedDiscreteDistrib}
\title{Truncated Analytical Expected Hessian (Discrete)}
\arguments{
\item{distrib}{A \code{TruncatedDiscreteDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of the parent's parameters.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
\deqn{\mathbb{E}\left[\dfrac{\partial^2 \ell_T}{\partial\theta_i\partial\theta_j}\right]
  = -\operatorname{Cov}_T(s_i, s_j)}
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
