% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated.R
\name{distrib_expected_hessian.ZeroInflatedDistrib}
\alias{distrib_expected_hessian.ZeroInflatedDistrib}
\title{Zero-Inflated Analytical Expected Hessian}
\arguments{
\item{distrib}{A \code{ZeroInflatedDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with the parent's parameters followed by \code{zi}.}
}
\value{
A list containing the vectors of expected second derivatives.
}
\description{
Expected Hessian (negative Fisher information) of the zero-inflated model, derived by
decomposing the expectation over \eqn{y=0} and \eqn{y>0} and using the parent's
expected Hessian.
}
\seealso{
\code{\link[=zero_inflated]{zero_inflated()}}
}
