% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta1_distrib.R
\name{distrib_grad_y.Beta1Distrib}
\alias{distrib_grad_y.Beta1Distrib}
\title{Beta Response Derivatives}
\arguments{
\item{distrib}{A \code{Beta1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Closed-form derivatives of the Beta log-density with respect to the response,
with \eqn{\alpha = \mu\phi}, \eqn{\beta = (1-\mu)\phi}:
\eqn{\partial \ell / \partial y = (\alpha-1)/y - (\beta-1)/(1-y)} and
\eqn{\partial^2 \ell / \partial y^2 = -(\alpha-1)/y^2 - (\beta-1)/(1-y)^2}.
}
\seealso{
\code{\link{beta1_distrib}}
}
