% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_distrib.R
\name{distrib_grad_y.ChisqDistrib}
\alias{distrib_grad_y.ChisqDistrib}
\title{Chi-Squared Response Gradient}
\arguments{
\item{distrib}{A \code{ChisqDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\deqn{\dfrac{\partial \ell}{\partial y} = \dfrac{\mu/2 - 1}{y} - \dfrac{1}{2}}
}
\seealso{
\code{\link{chisq_distrib}}
}
