% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_grad_y.ExponentialDistrib}
\alias{distrib_grad_y.ExponentialDistrib}
\title{Exponential Response Gradient}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\deqn{\dfrac{\partial \ell}{\partial y} = -\dfrac{1}{\mu}}
}
\seealso{
\code{\link[=exponential_distrib]{exponential_distrib()}}
}
