% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_distrib.R
\name{distrib_grad_y.Gamma1Distrib}
\alias{distrib_grad_y.Gamma1Distrib}
\title{Gamma Response Derivatives in Mean and Dispersion}
\arguments{
\item{distrib}{A \code{Gamma1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{phi}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial\ell/\partial y = (a-1)/y - b} at the implied
shape and rate.
}
\seealso{
\code{\link{gamma1_distrib}}
}
