% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian2_distrib.R
\name{distrib_grad_y.Gaussian2Distrib}
\alias{distrib_grad_y.Gaussian2Distrib}
\title{Gaussian Response Derivatives in Mean and Variance}
\arguments{
\item{distrib}{A \code{Gaussian2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{sigma2}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial\ell/\partial y = -(y-\mu)/v}.
}
\seealso{
\code{\link{gaussian2_distrib}}
}
