% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian3_distrib.R
\name{distrib_grad_y.Gaussian3Distrib}
\alias{distrib_grad_y.Gaussian3Distrib}
\title{Gaussian Response Derivatives in Mean and Precision}
\arguments{
\item{distrib}{A \code{Gaussian3Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{tau}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial\ell/\partial y = -\tau(y-\mu)}.
}
\seealso{
\code{\link{gaussian3_distrib}}
}
