% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss1_distrib.R
\name{distrib_grad_y.InvGauss1Distrib}
\alias{distrib_grad_y.InvGauss1Distrib}
\title{Inverse-Gaussian Response Derivatives}
\arguments{
\item{distrib}{An \code{InvGauss1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Closed-form derivatives of the Inverse-Gaussian log-density with respect to the
response:
\eqn{\partial \ell / \partial y = -\dfrac{3}{2y} - \dfrac{y^2 - \mu^2}{2\phi\mu^2 y^2}} and
\eqn{\partial^2 \ell / \partial y^2 = \dfrac{3}{2y^2} - \dfrac{1}{\phi y^3}}.
}
\seealso{
\code{\link[=invgauss1_distrib]{invgauss1_distrib()}}
}
