% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distrib.R
\name{distrib_gradient.BinomialDistrib}
\alias{distrib_gradient.BinomialDistrib}
\title{Binomial Analytical Gradient}
\arguments{
\item{distrib}{A \code{BinomialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of first derivatives.
}
\description{
Computes the analytical gradient (first derivative) of the Binomial log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{y - n\mu}{\mu(1-\mu)}}
}
\seealso{
\code{\link[=binomial_distrib]{binomial_distrib()}}
}
