% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distrib.R
\name{distrib_gradient.CauchyDistrib}
\alias{distrib_gradient.CauchyDistrib}
\title{Cauchy Analytical Gradient}
\arguments{
\item{distrib}{A \code{CauchyDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
Computes the analytical gradient (first derivatives) of the Cauchy log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{2(y-\mu)}{\sigma^2 + (y-\mu)^2}}
\deqn{\dfrac{\partial \ell}{\partial \sigma} = \dfrac{(y-\mu)^2 - \sigma^2}{\sigma(\sigma^2 + (y-\mu)^2)}}
}
\seealso{
\code{\link{cauchy_distrib}}
}
