% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_distrib.R
\name{distrib_gradient.GumbelDistrib}
\alias{distrib_gradient.GumbelDistrib}
\title{Gumbel Analytical Gradient}
\arguments{
\item{distrib}{A \code{GumbelDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{scale}{Either \code{"parameter"} or \code{"link"}.}

\item{...}{Unused.}
}
\value{
A named list of first derivatives.
}
\description{
Closed-form first derivatives of the Gumbel log-density, written in
\eqn{z = (y-\mu)/\sigma} and \eqn{w = e^{-z}}:
\deqn{\dfrac{\partial \ell}{\partial \mu} = \dfrac{1 - w}{\sigma},
      \qquad
      \dfrac{\partial \ell}{\partial \sigma}
        = \dfrac{z(1 - w) - 1}{\sigma}}
}
\seealso{
\code{\link[=gumbel_distrib]{gumbel_distrib()}}
}
