% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{distrib_gradient.ReparamContinuousDistrib}
\alias{distrib_gradient.ReparamContinuousDistrib}
\alias{reparam_gradient}
\title{Gradient of a Reparametrized Distribution}
\usage{
reparam_gradient(distrib, y, theta, scale = c("parameter", "link"), ...)
}
\arguments{
\item{distrib}{A reparametrized distribution.}

\item{y}{The response.}

\item{theta}{A named list of the new parameters.}

\item{scale}{Either \code{"parameter"} or \code{"link"}; handled by the generic.}

\item{...}{Unused.}
}
\value{
A named list of first derivatives.
}
\description{
The parent's score carried by the Jacobian of the map,
\eqn{\ell^{(a)} = \sum_i \ell^{(i)} \partial\theta_i/\partial\psi_a}.
}
\seealso{
\code{\link{reparametrize}}
}
\keyword{internal}
