% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{distrib_gradient.TruncatedDiscreteDistrib}
\alias{distrib_gradient.TruncatedDiscreteDistrib}
\title{Truncated Analytical Gradient (Discrete)}
\arguments{
\item{distrib}{A \code{TruncatedDiscreteDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of the parent's parameters.}
}
\value{
A list containing the vectors of first derivatives.
}
\description{
\deqn{\dfrac{\partial \ell_T}{\partial\theta_i} = s_i(y) - m_i, \qquad m_i = \mathbb{E}_T[s_i]}
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
