% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_hess_y.Beta2Distrib}
\alias{distrib_hess_y.Beta2Distrib}
\title{Beta Second Response Derivative in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial^2\ell/\partial y^2 = -(\alpha-1)/y^2 - (\beta-1)/(1-y)^2}.
}
\seealso{
\code{\link[=beta2_distrib]{beta2_distrib()}}
}
