% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian2_distrib.R
\name{distrib_hess_y.Gaussian2Distrib}
\alias{distrib_hess_y.Gaussian2Distrib}
\title{Gaussian Second Response Derivative in Mean and Variance}
\arguments{
\item{distrib}{A \code{Gaussian2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{sigma2}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\partial^2\ell/\partial y^2 = -1/v}, free of the response.
}
\seealso{
\code{\link{gaussian2_distrib}}
}
