% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_distrib.R
\name{distrib_hess_y.GumbelDistrib}
\alias{distrib_hess_y.GumbelDistrib}
\title{Gumbel Response Second Derivative}
\arguments{
\item{distrib}{A \code{GumbelDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\partial^2 \ell / \partial y^2 = -w/\sigma^2}.
}
\seealso{
\code{\link{gumbel_distrib}}
}
