% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss1_distrib.R
\name{distrib_hess_y.InvGauss1Distrib}
\alias{distrib_hess_y.InvGauss1Distrib}
\title{Inverse-Gaussian Response Second Derivative}
\arguments{
\item{distrib}{An \code{InvGauss1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = 3/(2y^2) - 1/(\phi y^3)}.
}
\seealso{
\code{\link[=invgauss1_distrib]{invgauss1_distrib()}}
}
