% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace2_distrib.R
\name{distrib_hess_y.Laplace2Distrib}
\alias{distrib_hess_y.Laplace2Distrib}
\title{Laplace Response Second Derivative in Location and Rate}
\arguments{
\item{distrib}{A \code{Laplace2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{lambda}.}
}
\value{
A numeric vector of zeros.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = 0} (almost everywhere).
}
\seealso{
\code{\link{laplace2_distrib}}
}
