% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_distrib.R
\name{distrib_hess_y.PseudoHuberDistrib}
\alias{distrib_hess_y.PseudoHuberDistrib}
\title{Pseudo-Huber Response Second Derivative}
\arguments{
\item{distrib}{A \code{PseudoHuberDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = -\nu/(\sigma^2 D^3)}, \eqn{D = \sqrt{\nu + ((y-\mu)/\sigma)^2}}.
}
\seealso{
\code{\link{pseudohuber_distrib}}
}
