% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal1_distrib.R
\name{distrib_hess_y.SkewNormal1Distrib}
\alias{distrib_hess_y.SkewNormal1Distrib}
\title{Skew Normal Response Second Derivative}
\arguments{
\item{distrib}{A \code{SkewNormal1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{mu}, \code{sigma} and \code{alpha}.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\partial^2 \ell / \partial y^2 = (\alpha^2 R' - 1)/\sigma^2}.
}
\seealso{
\code{\link{skewnormal1_distrib}}
}
