% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal2_distrib.R
\name{distrib_hess_y.SkewNormal2Distrib}
\alias{distrib_hess_y.SkewNormal2Distrib}
\title{Skew Normal Second Response Derivative in the Centered Parametrization}
\arguments{
\item{distrib}{A \code{SkewNormal2Distrib} object.}

\item{y}{The response.}

\item{theta}{A list with \code{mu}, \code{sigma} and \code{gamma1}.}

\item{...}{Passed on.}
}
\value{
A numeric vector.
}
\description{
The parent's, unchanged.
}
\seealso{
\code{\link{skewnormal2_distrib}}
}
