% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewt_distrib.R
\name{distrib_hess_y.SkewTDistrib}
\alias{distrib_hess_y.SkewTDistrib}
\title{Skew t Response Second Derivative}
\arguments{
\item{distrib}{A \code{SkewTDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{mu}, \code{sigma}, \code{alpha} and \code{nu}.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\partial^2 \ell / \partial y^2 = D'/\sigma^2}.
}
\seealso{
\code{\link[=skewt_distrib]{skewt_distrib()}}
}
