% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t1_distrib.R
\name{distrib_hess_y.StudentT1Distrib}
\alias{distrib_hess_y.StudentT1Distrib}
\title{Student's t Response Second Derivative}
\arguments{
\item{distrib}{A \code{StudentT1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}
}
\value{
A numeric vector.
}
\description{
Closed-form \eqn{\partial^2 \ell / \partial y^2 = (\nu+1)(r^2 - \nu\sigma^2)/(\nu\sigma^2 + r^2)^2}, \eqn{r = y - \mu}.
}
\seealso{
\code{\link{student_t1_distrib}}
}
