% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{distrib_hess_y.TruncatedContinuousDistrib}
\alias{distrib_hess_y.TruncatedContinuousDistrib}
\title{Truncated Continuous Response Hessian}
\arguments{
\item{distrib}{A \code{TruncatedContinuousDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of the parent's parameters.}
}
\value{
A numeric vector.
}
\description{
As \code{\link[=distrib_grad_y.TruncatedContinuousDistrib]{distrib_grad_y()}}, at second order.
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
