% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1_distrib.R
\name{distrib_hess_y.Weibull1Distrib}
\alias{distrib_hess_y.Weibull1Distrib}
\title{Weibull Response Second Derivative}
\arguments{
\item{distrib}{A \code{Weibull1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector.
}
\description{
Closed form:
\eqn{\partial^2 \ell / \partial y^2 = -(\sigma - 1)(1 + \sigma u)/y^2}.
}
\seealso{
\code{\link[=weibull1_distrib]{weibull1_distrib()}}
}
