% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distrib.R
\name{distrib_hessian.BernoulliDistrib}
\alias{distrib_hessian.BernoulliDistrib}
\title{Bernoulli Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{BernoulliDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A list containing the vector of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivative) of the Bernoulli log-probability
with respect to the parameter \eqn{\mu}.

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = -\dfrac{y}{\mu^2} - \dfrac{1-y}{(1-\mu)^2}}
}
\seealso{
\code{\link{bernoulli_distrib}}
}
