% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distrib.R
\name{distrib_hessian.CauchyDistrib}
\alias{distrib_hessian.CauchyDistrib}
\title{Cauchy Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{CauchyDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivatives) of the Cauchy log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = \dfrac{2(y-\mu)^2 - 2\sigma^2}{(\sigma^2 + (y-\mu)^2)^2}}
\deqn{\dfrac{\partial^2 \ell}{\partial \sigma^2} = \dfrac{\sigma^4 - 4\sigma^2 (y-\mu)^2 - (y-\mu)^4}{\sigma^2(\sigma^2 + (y-\mu)^2)^2}}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma} = -\dfrac{4\sigma (y-\mu)}{(\sigma^2 + (y-\mu)^2)^2}}
}
\seealso{
\code{\link{cauchy_distrib}}
}
