% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_derivatives.R
\name{distrib_hessian.distrib}
\alias{distrib_hessian.distrib}
\title{Default Numerical Hessian for \code{distrib} Objects}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of parameters.}
}
\value{
A named list of Hessian component vectors in \code{\link[=hess_names]{hess_names()}} order.
}
\description{
Fallback method: distributions that do not implement an analytical Hessian get
one computed by finite differences of the log-density
(see \code{\link[=numerical_hessian]{numerical_hessian()}}).
}
\keyword{internal}
