% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_hessian.ExponentialDistrib}
\alias{distrib_hessian.ExponentialDistrib}
\title{Exponential Analytical Observed Hessian}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{scale}{Either \code{"parameter"} or \code{"link"}; handled by the generic.}

\item{...}{Unused.}
}
\value{
A named list with the \code{mu_mu} component.
}
\description{
\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = \dfrac{1}{\mu^2} - \dfrac{2y}{\mu^3}}
}
\seealso{
\code{\link[=exponential_distrib]{exponential_distrib()}}
}
