% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian1_distrib.R
\name{distrib_hessian.Gaussian1Distrib}
\alias{distrib_hessian.Gaussian1Distrib}
\title{Gaussian Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{Gaussian1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivatives) of the Gaussian log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma}.

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = -\dfrac{1}{\sigma^2}}
\deqn{\dfrac{\partial^2 \ell}{\partial \sigma^2} = \dfrac{\sigma^2 - 3(y - \mu)^2}{\sigma^4}}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma} = -\dfrac{2(y - \mu)}{\sigma^3}}
}
\seealso{
\code{\link{gaussian1_distrib}}
}
