% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengamma1_distrib.R
\name{distrib_hessian.GenGamma1Distrib}
\alias{distrib_hessian.GenGamma1Distrib}
\title{Generalised Gamma Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{GenGamma1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{a}, \code{d} and \code{p}.}

\item{scale}{Either \code{"parameter"} or \code{"link"}; handled by the generic.}

\item{...}{Unused.}
}
\value{
A named list of second-derivative components.
}
\description{
The second derivatives of the same expressions. The mixed
\eqn{a}--\eqn{d} component is \eqn{-1/a}, free of the data, the scale and
the first shape entering the log-density through \eqn{-d\log a} alone.
}
\seealso{
\code{\link{gengamma1_distrib}}
}
