% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_distrib.R
\name{distrib_hessian.GumbelDistrib}
\alias{distrib_hessian.GumbelDistrib}
\title{Gumbel Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{GumbelDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{scale}{Either \code{"parameter"} or \code{"link"}.}

\item{...}{Unused.}
}
\value{
A named list of second derivatives.
}
\description{
Closed-form second derivatives of the Gumbel log-density:
\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = -\dfrac{w}{\sigma^2},
      \qquad
      \dfrac{\partial^2 \ell}{\partial \mu \, \partial \sigma}
        = -\dfrac{1 - w + z w}{\sigma^2},}
\deqn{\dfrac{\partial^2 \ell}{\partial \sigma^2}
        = \dfrac{1 - 2z + 2zw - z^2 w}{\sigma^2}.}
}
\seealso{
\code{\link{gumbel_distrib}}
}
