% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distrib.R
\name{distrib_hessian.LognormalDistrib}
\alias{distrib_hessian.LognormalDistrib}
\title{Lognormal Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{LognormalDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian (second derivatives) of the Lognormal log-density
with respect to the parameters \eqn{\mu} and \eqn{\sigma^2}.
}
\seealso{
\code{\link{lognormal_distrib}}
}
