% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_distrib.R
\name{distrib_hessian.PseudoHuberDistrib}
\alias{distrib_hessian.PseudoHuberDistrib}
\title{Pseudo-Huber Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{PseudoHuberDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
Computes the analytical observed Hessian of the Pseudo-Huber log-density.
Let \eqn{r = y - \mu}, \eqn{D = \sqrt{\nu + (r/\sigma)^2}},
\eqn{R_1 = K_1'(\sqrt{\nu})/K_1(\sqrt{\nu})} and \eqn{R_2 = K_1''(\sqrt{\nu})/K_1(\sqrt{\nu})}:

\deqn{\dfrac{\partial^2 \ell}{\partial \mu^2} = -\dfrac{\nu}{\sigma^2 D^3}}
\deqn{\dfrac{\partial^2 \ell}{\partial \sigma^2} = \dfrac{\sigma^4 - 3\sigma^2 r^2 D^{-1} + r^4 D^{-3}}{\sigma^6}}
\deqn{\dfrac{\partial^2 \ell}{\partial \nu^2} = \dfrac{1}{4D^3} + \dfrac{1}{2\nu^2} + \dfrac{1}{4}\left(\dfrac{R_1}{\nu^{3/2}} + \dfrac{R_1^2}{\nu} - \dfrac{R_2}{\nu}\right)}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \partial \sigma} = \dfrac{-2\nu\sigma^2 r - r^3}{\sigma^2(\nu\sigma^2 + r^2)^{3/2}}}
\deqn{\dfrac{\partial^2 \ell}{\partial \mu \partial \nu} = -\dfrac{r}{2\sigma^2 D^3}}
\deqn{\dfrac{\partial^2 \ell}{\partial \sigma \partial \nu} = -\dfrac{r^2}{2\sigma^3 D^3}}
}
\seealso{
\code{\link[=pseudohuber_distrib]{pseudohuber_distrib()}}
}
