% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{distrib_hessian.TransformedDistrib}
\alias{distrib_hessian.TransformedDistrib}
\title{Transformed Analytical Observed Hessian}
\arguments{
\item{distrib}{A \code{TransformedDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list of the parent's parameters.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
The observed Hessian of the transformed model equals the parent's observed Hessian
evaluated at \eqn{x = g^{-1}(y)}.
}
\seealso{
\code{\link[=transformation]{transformation()}}
}
