% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{distrib_hessian.TruncatedDiscreteDistrib}
\alias{distrib_hessian.TruncatedDiscreteDistrib}
\title{Truncated Analytical Observed Hessian (Discrete)}
\arguments{
\item{distrib}{A \code{TruncatedDiscreteDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of the parent's parameters.}
}
\value{
A list containing the vectors of second derivatives.
}
\description{
\deqn{\dfrac{\partial^2 \ell_T}{\partial\theta_i\partial\theta_j} = H_{ij}(y) - M_{ij} + m_i m_j}
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
