% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_pdf.Beta2Distrib}
\alias{distrib_pdf.Beta2Distrib}
\title{Beta Density in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{y}{A numeric vector of observations in \eqn{(0, 1)}.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector.
}
\description{
\deqn{f(y) = \dfrac{y^{\alpha-1}(1-y)^{\beta-1}}{B(\alpha, \beta)}}
}
\seealso{
\code{\link[=beta2_distrib]{beta2_distrib()}}
}
