% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom1_distrib.R
\name{distrib_pdf.BetaBinom1Distrib}
\alias{distrib_pdf.BetaBinom1Distrib}
\title{Beta-Binomial Probability Mass Function}
\arguments{
\item{distrib}{A \code{BetaBinom1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{mu} and \code{sigma}.}

\item{log}{Logical; if \code{TRUE}, returns the log-probability.}
}
\value{
A numeric vector of probability values.
}
\description{
\deqn{P(Y = y) = \binom{n}{y}
      \dfrac{B(y + \alpha,\; n - y + \beta)}{B(\alpha, \beta)}, \qquad
      \alpha = \dfrac{\mu}{\sigma}, \quad \beta = \dfrac{1-\mu}{\sigma}}
}
\seealso{
\code{\link[=betabinom1_distrib]{betabinom1_distrib()}}
}
