% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom2_distrib.R
\name{distrib_pdf.BetaBinom2Distrib}
\alias{distrib_pdf.BetaBinom2Distrib}
\title{Beta-Binomial Mass Function in Its Shapes}
\arguments{
\item{distrib}{A \code{BetaBinom2Distrib} object.}

\item{y}{A numeric vector of counts.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{log}{Logical; if \code{TRUE}, returns the log-probability.}
}
\value{
A numeric vector.
}
\description{
\deqn{P(Y = y) = \binom{n}{y}\dfrac{B(y+\alpha, n-y+\beta)}{B(\alpha,\beta)}}
}
\seealso{
\code{\link{betabinom2_distrib}}
}
