% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_distrib.R
\name{distrib_pdf.ChisqDistrib}
\alias{distrib_pdf.ChisqDistrib}
\title{Chi-Squared Density}
\arguments{
\item{distrib}{A \code{ChisqDistrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector of density values.
}
\description{
\deqn{f(y; \mu) = \dfrac{y^{\mu/2 - 1} e^{-y/2}}{2^{\mu/2}\Gamma(\mu/2)},
      \qquad y > 0}
}
\seealso{
\code{\link{chisq_distrib}}
}
