% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian2_distrib.R
\name{distrib_pdf.Gaussian2Distrib}
\alias{distrib_pdf.Gaussian2Distrib}
\title{Gaussian Density in Mean and Variance}
\arguments{
\item{distrib}{A \code{Gaussian2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{sigma2}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector.
}
\description{
\deqn{f(y) = \dfrac{1}{\sqrt{2\pi\sigma^2}}
      \exp\left\{-\dfrac{(y-\mu)^2}{2\sigma^2}\right\}}
}
\seealso{
\code{\link{gaussian2_distrib}}
}
