% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian3_distrib.R
\name{distrib_pdf.Gaussian3Distrib}
\alias{distrib_pdf.Gaussian3Distrib}
\title{Gaussian Density in Mean and Precision}
\arguments{
\item{distrib}{A \code{Gaussian3Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{tau}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector.
}
\description{
\deqn{f(y) = \sqrt{\dfrac{\tau}{2\pi}}
      \exp\left\{-\dfrac{\tau(y-\mu)^2}{2}\right\}}
}
\seealso{
\code{\link{gaussian3_distrib}}
}
