% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengamma1_distrib.R
\name{distrib_pdf.GenGamma1Distrib}
\alias{distrib_pdf.GenGamma1Distrib}
\title{Generalized Gamma Density}
\arguments{
\item{distrib}{A \code{GenGamma1Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing \code{a}, \code{d} and \code{p}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector of density values.
}
\description{
\deqn{f(y) = \dfrac{p}{a^{d}\,\Gamma(d/p)}\, y^{d-1} e^{-(y/a)^{p}},
      \qquad y > 0}
}
\seealso{
\code{\link{gengamma1_distrib}}
}
