% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss2_distrib.R
\name{distrib_pdf.InvGauss2Distrib}
\alias{distrib_pdf.InvGauss2Distrib}
\title{Inverse Gaussian Density in Mean and Shape}
\arguments{
\item{distrib}{An \code{InvGauss2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list with \code{mu} and \code{lambda}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector.
}
\description{
\deqn{f(y) = \sqrt{\dfrac{\lambda}{2\pi y^3}}
      \exp\left\{-\dfrac{\lambda(y-\mu)^2}{2\mu^2 y}\right\}}
}
\seealso{
\code{\link{invgauss2_distrib}}
}
