% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace2_distrib.R
\name{distrib_pdf.Laplace2Distrib}
\alias{distrib_pdf.Laplace2Distrib}
\title{Laplace Density in Location and Rate}
\arguments{
\item{distrib}{A \code{Laplace2Distrib} object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A list containing the parameters \code{mu} and \code{lambda}.}

\item{log}{Logical; if \code{TRUE}, returns the log-density.}
}
\value{
A numeric vector of density values.
}
\description{
\deqn{f(y; \mu, \lambda) = \dfrac{\lambda}{2} \exp\left(-\lambda|y-\mu|\right)}
}
\seealso{
\code{\link[=laplace2_distrib]{laplace2_distrib()}}
}
