% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_quantile.Beta2Distrib}
\alias{distrib_quantile.Beta2Distrib}
\title{Beta Quantile Function in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list with \code{alpha} and \code{beta}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is a log-probability.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The inverse of the incomplete beta function.
}
\seealso{
\code{\link[=beta2_distrib]{beta2_distrib()}}
}
