% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distrib.R
\name{distrib_quantile.BetaDistrib}
\alias{distrib_quantile.BetaDistrib}
\title{Beta Quantile Function}
\arguments{
\item{distrib}{A \code{BetaDistrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le p)}, otherwise \eqn{P(Y > p)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the quantile function (inverse CDF) for the Beta distribution.
}
\seealso{
\code{\link{beta_distrib}}
}
