% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distrib.R
\name{distrib_quantile.BinomialDistrib}
\alias{distrib_quantile.BinomialDistrib}
\title{Binomial Quantile Function}
\arguments{
\item{distrib}{A \code{BinomialDistrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le p)}, otherwise \eqn{P(Y > p)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\description{
Computes the quantile function (inverse CDF) for the Binomial distribution.
}
\seealso{
\code{\link{binomial_distrib}}
}
