% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_distrib.R
\name{distrib_quantile.ChisqDistrib}
\alias{distrib_quantile.ChisqDistrib}
\title{Chi-Squared Quantile Function}
\arguments{
\item{distrib}{A \code{ChisqDistrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameter \code{mu}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), \code{p} is \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is given as its logarithm.}
}
\value{
A numeric vector of quantiles.
}
\description{
The inverse of \code{\link[stats]{pchisq}}.
}
\seealso{
\code{\link{chisq_distrib}}
}
