% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian2_distrib.R
\name{distrib_quantile.Gaussian2Distrib}
\alias{distrib_quantile.Gaussian2Distrib}
\title{Gaussian Quantile Function in Mean and Variance}
\arguments{
\item{distrib}{A \code{Gaussian2Distrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list with \code{mu} and \code{sigma2}.}

\item{lower.tail}{Logical; if \code{TRUE}, probabilities are \eqn{P(Y \le q)}.}

\item{log.p}{Logical; if \code{TRUE}, \code{p} is given as a log-probability.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
The normal quantile function.
}
\seealso{
\code{\link[=gaussian2_distrib]{gaussian2_distrib()}}
}
