% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace2_distrib.R
\name{distrib_quantile.Laplace2Distrib}
\alias{distrib_quantile.Laplace2Distrib}
\title{Laplace Quantile Function in Location and Rate}
\arguments{
\item{distrib}{A \code{Laplace2Distrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameters \code{mu} and \code{lambda}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le p)}, otherwise \eqn{P(Y > p)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\value{
A numeric vector of quantiles.
}
\description{
\deqn{Q(p; \mu, \lambda) = \mu - \dfrac{1}{\lambda}\,\mathrm{sign}(p - \tfrac{1}{2})\,\log\left(1 - 2\left|p - \tfrac{1}{2}\right|\right)}
}
\seealso{
\code{\link{laplace2_distrib}}
}
