% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distrib.R
\name{distrib_quantile.LogisticDistrib}
\alias{distrib_quantile.LogisticDistrib}
\title{Logistic Quantile Function}
\arguments{
\item{distrib}{A \code{LogisticDistrib} object.}

\item{p}{A numeric vector of probabilities.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(Y \le p)}, otherwise \eqn{P(Y > p)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}
}
\value{
A numeric vector of quantiles.
}
\description{
Computes the quantile function (inverse CDF) for the Logistic distribution:
\deqn{Q(p; \mu, \sigma) = \mu + \sigma \log\left(\dfrac{p}{1-p}\right)}
}
\seealso{
\code{\link[=logistic_distrib]{logistic_distrib()}}
}
