% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distrib_quantile}
\alias{distrib_quantile}
\title{Quantile Function}
\usage{
distrib_quantile(distrib, ...)
}
\arguments{
\item{distrib}{A distribution object inheriting from the \code{distrib} class.}

\item{...}{Additional arguments passed to the specific method (e.g., \code{p}, \code{theta}).}
}
\description{
Evaluates the quantile function for a given distribution.
}
