% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distrib.R
\name{distrib_rng.BernoulliDistrib}
\alias{distrib_rng.BernoulliDistrib}
\title{Bernoulli Random Number Generator}
\arguments{
\item{distrib}{A \code{BernoulliDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\description{
Generates random numbers from the Bernoulli distribution.
}
\seealso{
\code{\link{bernoulli_distrib}}
}
