% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_distrib.R
\name{distrib_rng.Beta2Distrib}
\alias{distrib_rng.Beta2Distrib}
\title{Beta Random Generation in Its Shapes}
\arguments{
\item{distrib}{A \code{Beta2Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{alpha} and \code{beta}.}
}
\value{
A numeric vector.
}
\description{
Delegates to \code{\link[stats:Beta]{stats::rbeta()}}.
}
\seealso{
\code{\link[=beta2_distrib]{beta2_distrib()}}
}
