% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distrib.R
\name{distrib_rng.BinomialDistrib}
\alias{distrib_rng.BinomialDistrib}
\title{Binomial Random Number Generator}
\arguments{
\item{distrib}{A \code{BinomialDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Binomial distribution.
}
\seealso{
\code{\link[=binomial_distrib]{binomial_distrib()}}
}
