% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distrib.R
\name{distrib_rng.CauchyDistrib}
\alias{distrib_rng.CauchyDistrib}
\title{Cauchy Random Number Generator}
\arguments{
\item{distrib}{A \code{CauchyDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Cauchy distribution.
}
\seealso{
\code{\link{cauchy_distrib}}
}
