% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_distrib.R
\name{distrib_rng.ChisqDistrib}
\alias{distrib_rng.ChisqDistrib}
\title{Chi-Squared Random Generation}
\arguments{
\item{distrib}{A \code{ChisqDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draws through \code{\link[stats]{rchisq}}.
}
\seealso{
\code{\link{chisq_distrib}}
}
