% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{distrib_rng.discrete_distrib}
\alias{distrib_rng.discrete_distrib}
\title{Default Numerical RNG for Discrete Distributions}
\arguments{
\item{distrib}{An object inheriting from class \code{"discrete_distrib"}.}

\item{n}{Number of observations to generate.}

\item{theta}{A named list of parameters.}
}
\description{
Fallback method: discrete distributions that do not implement a native RNG
generate draws by inverse transform sampling on the cumulative pmf table.
}
\keyword{internal}
