% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet_distrib.R
\name{distrib_rng.EnetDistrib}
\alias{distrib_rng.EnetDistrib}
\title{Elastic-Net Random Generation}
\arguments{
\item{distrib}{An \code{EnetDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing \code{mu}, \code{lambda} and \code{alpha}.}
}
\value{
A numeric vector of draws.
}
\description{
Inverse transform on the closed-form quantile function.
}
\seealso{
\code{\link{enet_distrib}}
}
