% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distrib.R
\name{distrib_rng.ExponentialDistrib}
\alias{distrib_rng.ExponentialDistrib}
\title{Exponential Random Generation}
\arguments{
\item{distrib}{An \code{ExponentialDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draws from the exponential distribution through
\code{\link[stats:Exponential]{stats::rexp()}}.
}
\seealso{
\code{\link[=exponential_distrib]{exponential_distrib()}}
}
