% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_distrib.R
\name{distrib_rng.Gamma1Distrib}
\alias{distrib_rng.Gamma1Distrib}
\title{Gamma Random Generation in Mean and Dispersion}
\arguments{
\item{distrib}{A \code{Gamma1Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{mu} and \code{phi}.}
}
\value{
A numeric vector.
}
\description{
Delegates to \code{\link[stats]{rgamma}}.
}
\seealso{
\code{\link{gamma1_distrib}}
}
