% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distrib.R
\name{distrib_rng.GammaDistrib}
\alias{distrib_rng.GammaDistrib}
\title{Gamma Random Number Generator}
\arguments{
\item{distrib}{A \code{GammaDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma2}.}
}
\description{
Generates random numbers from the Gamma distribution.
}
\seealso{
\code{\link{gamma_distrib}}
}
