% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian3_distrib.R
\name{distrib_rng.Gaussian3Distrib}
\alias{distrib_rng.Gaussian3Distrib}
\title{Gaussian Random Generation in Mean and Precision}
\arguments{
\item{distrib}{A \code{Gaussian3Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{mu} and \code{tau}.}
}
\value{
A numeric vector.
}
\description{
Delegates to \code{\link[stats:Normal]{stats::rnorm()}}.
}
\seealso{
\code{\link[=gaussian3_distrib]{gaussian3_distrib()}}
}
