% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distrib.R
\name{distrib_rng.GeometricDistrib}
\alias{distrib_rng.GeometricDistrib}
\title{Geometric Random Generation}
\arguments{
\item{distrib}{A \code{GeometricDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draws through \code{\link[stats]{rgeom}} at
\eqn{p = 1/(1+\mu)}.
}
\seealso{
\code{\link{geometric_distrib}}
}
