% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_distrib.R
\name{distrib_rng.GPDDistrib}
\alias{distrib_rng.GPDDistrib}
\title{Generalized Pareto Random Generation}
\arguments{
\item{distrib}{A \code{GPDDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list containing \code{sigma} and \code{xi}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Inverse transform, the quantile function being elementary.
}
\seealso{
\code{\link{gpd_distrib}}
}
