% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss1_distrib.R
\name{distrib_rng.InvGauss1Distrib}
\alias{distrib_rng.InvGauss1Distrib}
\title{Inverse-Gaussian Random Number Generator}
\arguments{
\item{distrib}{An \code{InvGauss1Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{phi}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Inverse-Gaussian distribution.
}
\seealso{
\code{\link{invgauss1_distrib}}
}
