% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgauss2_distrib.R
\name{distrib_rng.InvGauss2Distrib}
\alias{distrib_rng.InvGauss2Distrib}
\title{Inverse Gaussian Random Generation in Mean and Shape}
\arguments{
\item{distrib}{An \code{InvGauss2Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{mu} and \code{lambda}.}
}
\value{
A numeric vector.
}
\description{
Delegates to \code{\link[statmod:invgauss]{statmod::rinvgauss()}}.
}
\seealso{
\code{\link[=invgauss2_distrib]{invgauss2_distrib()}}
}
