% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace2_distrib.R
\name{distrib_rng.Laplace2Distrib}
\alias{distrib_rng.Laplace2Distrib}
\title{Laplace Random Generation in Location and Rate}
\arguments{
\item{distrib}{A \code{Laplace2Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{lambda}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates draws by inverse-transform sampling.
}
\seealso{
\code{\link{laplace2_distrib}}
}
