% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distrib.R
\name{distrib_rng.LaplaceDistrib}
\alias{distrib_rng.LaplaceDistrib}
\title{Laplace Random Number Generator}
\arguments{
\item{distrib}{A \code{LaplaceDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{b}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Laplace distribution by inverse-transform
sampling.
}
\seealso{
\code{\link{laplace_distrib}}
}
