% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distrib.R
\name{distrib_rng.MultinomialDistrib}
\alias{distrib_rng.MultinomialDistrib}
\title{Multinomial Random Generation}
\arguments{
\item{distrib}{A \code{MultinomialDistrib} object.}

\item{n}{The number of draws.}

\item{theta}{A named list of parameters.}
}
\value{
A matrix with \code{n} rows.
}
\description{
\code{\link[stats]{rmultinom}}, transposed so that one row is
one observation.
}
\seealso{
\code{\link{multinomial_distrib}}
}
