% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin2_distrib.R
\name{distrib_rng.NegBin2Distrib}
\alias{distrib_rng.NegBin2Distrib}
\title{Negative Binomial Random Number Generator}
\arguments{
\item{distrib}{A \code{NegBin2Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{theta}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Negative Binomial distribution.
}
\seealso{
\code{\link[=negbin2_distrib]{negbin2_distrib()}}
}
