% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_distrib.R
\name{distrib_rng.NegBinDistrib}
\alias{distrib_rng.NegBinDistrib}
\title{Negative Binomial Random Number Generator}
\arguments{
\item{distrib}{A \code{NegBinDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{theta}.}
}
\description{
Generates random numbers from the Negative Binomial distribution.
}
\seealso{
\code{\link{negbin_distrib}}
}
