% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distrib.R
\name{distrib_rng.PoissonDistrib}
\alias{distrib_rng.PoissonDistrib}
\title{Poisson Random Number Generator}
\arguments{
\item{distrib}{A \code{PoissonDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameter \code{mu}.}
}
\description{
Generates random numbers from the Poisson distribution.
}
\seealso{
\code{\link{poisson_distrib}}
}
