% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_distrib.R
\name{distrib_rng.PseudoHuberDistrib}
\alias{distrib_rng.PseudoHuberDistrib}
\title{Pseudo-Huber Random Number Generator}
\arguments{
\item{distrib}{A \code{PseudoHuberDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma} and \code{nu}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Pseudo-Huber distribution via inverse transform
sampling (numerical quantile function applied to uniform draws).
}
\seealso{
\code{\link[=pseudohuber_distrib]{pseudohuber_distrib()}}
}
