% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distrib_rng}
\alias{distrib_rng}
\title{Random Number Generator}
\usage{
distrib_rng(distrib, ...)
}
\arguments{
\item{distrib}{A distribution object inheriting from the \code{distrib} class.}

\item{...}{Additional arguments passed to the specific method (e.g., \code{n}, \code{theta}).}
}
\description{
Generates random variates from the given distribution.
}
