% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t1_distrib.R
\name{distrib_rng.StudentT1Distrib}
\alias{distrib_rng.StudentT1Distrib}
\title{Student's t Random Number Generator}
\arguments{
\item{distrib}{A \code{StudentT1Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu}, \code{sigma}, and \code{nu}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers from the Student's t distribution.
}
\seealso{
\code{\link{student_t1_distrib}}
}
