% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{distrib_rng.TruncatedDiscreteDistrib}
\alias{distrib_rng.TruncatedDiscreteDistrib}
\title{Truncated Random Number Generator (Discrete)}
\arguments{
\item{distrib}{A \code{TruncatedDiscreteDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A named list of the parent's parameters.}
}
\value{
A numeric vector of random draws.
}
\description{
Inverse transform sampling on the parent, exact for a discrete cdf.
}
\seealso{
\code{\link[=truncated]{truncated()}}
}
