% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmises2_distrib.R
\name{distrib_rng.VonMises2Distrib}
\alias{distrib_rng.VonMises2Distrib}
\title{von Mises Random Generation in the Resultant Length}
\arguments{
\item{distrib}{A \code{VonMises2Distrib} object.}

\item{n}{The number of draws.}

\item{theta}{A list with \code{mu} and \code{rho}.}
}
\value{
A numeric vector of angles.
}
\description{
Delegates to the concentration parametrization.
}
\seealso{
\code{\link{vonmises2_distrib}}
}
