% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1_distrib.R
\name{distrib_rng.Weibull1Distrib}
\alias{distrib_rng.Weibull1Distrib}
\title{Weibull Random Number Generator}
\arguments{
\item{distrib}{A \code{Weibull1Distrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list containing the parameters \code{mu} and \code{sigma}.}
}
\value{
A numeric vector of random draws.
}
\description{
Generates random numbers by inverse transform, which is exact here because
the quantile function is elementary.
}
\seealso{
\code{\link[=weibull1_distrib]{weibull1_distrib()}}
}
