% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated.R
\name{distrib_rng.ZeroInflatedDistrib}
\alias{distrib_rng.ZeroInflatedDistrib}
\title{Zero-Inflated Random Number Generator}
\arguments{
\item{distrib}{A \code{ZeroInflatedDistrib} object.}

\item{n}{Number of observations to generate.}

\item{theta}{A list with the parent's parameters followed by \code{zi}.}
}
\description{
Draws from the wrapped distribution, then replaces a Bernoulli(\eqn{\zeta}) fraction
of the draws with structural zeros.
}
\seealso{
\code{\link{zero_inflated}}
}
