% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributions}
\alias{distributions}
\alias{Distribution}
\alias{DistributionDiscrete}
\alias{DistributionContinuous}
\title{Distribution classes}
\usage{
Distribution(name = character(0), support = Real())

DistributionDiscrete(name = character(0), support = Int())

DistributionContinuous(name = character(0), support = Real())
}
\arguments{
\item{name}{Name of the distribution.}

\item{support}{Support of the distribution.}
}
\description{
These classes are abstract and provide the base classes for all individual distributions in the package.
These classes are not intended for direct use, unless you want to implement a custom distribution class.

Look below ("see also") for a list of all available distributions
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
\keyword{internal}
