% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributions}
\alias{distributions}
\alias{Distribution}
\alias{DistributionDiscrete}
\alias{DistributionContinuous}
\title{Distribution classes}
\usage{
Distribution(name = character(0), support = Real())

DistributionDiscrete(name = character(0), support = Int())

DistributionContinuous(name = character(0), support = Real())
}
\arguments{
\item{name}{Name of the distribution.}

\item{support}{Support of the distribution.}
}
\description{
These classes are abstract and provide the base classes for all individual distributions in the package.
These classes are not intended for direct use, unless you want to implement a custom distribution class.

Look below ("see also") for a list of all available distributions
}
\seealso{
Other distributions: 
\code{\link{amoroso}()},
\code{\link{beta}()},
\code{\link{beta_prime}()},
\code{\link{binomial}()},
\code{\link{cauchy}()},
\code{\link{chi_squared}()},
\code{\link{exponential}()},
\code{\link{f-distribution}},
\code{\link{frechet}()},
\code{\link{gamma}()},
\code{\link{gompertz}()},
\code{\link{gumbel}()},
\code{\link{inverse_gamma}()},
\code{\link{laplace}()},
\code{\link{log_logistic}()},
\code{\link{log_normal}()},
\code{\link{logistic}()},
\code{\link{negative_binomial}()},
\code{\link{normal}()},
\code{\link{pareto}()},
\code{\link{shifted}()},
\code{\link{skew}},
\code{\link{skewed_generalized_t}()},
\code{\link{stretched_beta}()},
\code{\link{symmetric_generalized_normal}()},
\code{\link{t-distribution}},
\code{\link{triangular}()},
\code{\link{uniform}()},
\code{\link{wald}()},
\code{\link{weibull}()}
}
\concept{distributions}
\keyword{internal}
