% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{exp_transform}
\alias{exp_transform}
\title{Exponential Transformation}
\usage{
exp_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = e^X}: maps the real line to \eqn{(0, \infty)}.
Inverse \eqn{X = \log(Y)}, Jacobian \eqn{|J| = 1/Y}.
}
\examples{
d <- transformation(gaussian_distrib(), exp_transform())
distrib_pdf(d, 1, list(mu = 0, sigma = 1))

}
