% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{expit_transform}
\alias{expit_transform}
\title{Expit (Sigmoid) Transformation}
\usage{
expit_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \text{plogis}(X)}: maps the real line to \eqn{(0, 1)}.
Inverse \eqn{X = \text{logit}(Y)}, Jacobian \eqn{|J| = 1/(Y(1-Y))}.
}
\examples{
d <- transformation(gaussian1_distrib(), expit_transform())
distrib_pdf(d, 0.5, list(mu = 0, sigma = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{bc_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{logit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
