% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distrib.R
\name{fit_score}
\alias{fit_score}
\title{Summed Score on the Link Scale}
\usage{
fit_score(distrib, y, theta, threads = 1L)
}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of parameters on the natural scale.}
}
\value{
A numeric vector of length \code{length(distrib@params)}.
}
\description{
The gradient of the total log-likelihood with respect to \eqn{\eta}, summed
over observations.
}
\seealso{
\code{\link{fit_hess_matrix}}
}
\keyword{internal}
