% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate_random_theta}
\alias{generate_random_theta}
\title{Generate Random Parameters}
\usage{
generate_random_theta(distrib, ...)
}
\arguments{
\item{distrib}{A distribution object inheriting from the \code{distrib} class.}

\item{...}{Additional arguments passed to the specific method.}
}
\description{
Generates sensible random parameters for a distribution based on its mathematical domain.
}
