% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Reciprocal (Inverse) Transformation}
\usage{
inverse_transform()
}
\value{
A \code{\link[=transformer]{transformer()}} object.
}
\description{
Transformer for \eqn{Y = 1/X} (monotonically decreasing on a support not
containing 0). Inverse \eqn{X = 1/Y}, Jacobian \eqn{|J| = 1/Y^2}.
}
\examples{
d <- transformation(gamma2_distrib(), inverse_transform())
distrib_pdf(d, 1, list(mu = 2, sigma2 = 1))

}
\seealso{
\code{\link[=log_transform]{log_transform()}}, \code{\link[=exp_transform]{exp_transform()}}, \code{\link[=sqrt_transform]{sqrt_transform()}}, \code{\link[=power_transform]{power_transform()}}, \code{\link[=bc_transform]{bc_transform()}}, \code{\link[=yj_transform]{yj_transform()}}, \code{\link[=softplus_transform]{softplus_transform()}}, \code{\link[=asinh_transform]{asinh_transform()}}, \code{\link[=logit_transform]{logit_transform()}}, \code{\link[=expit_transform]{expit_transform()}}, \code{\link[=affine_transform]{affine_transform()}}, \code{\link[=transformation]{transformation()}}
}
