% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Reciprocal (Inverse) Transformation}
\usage{
inverse_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = 1/X} (monotonically decreasing on a support not
containing 0). Inverse \eqn{X = 1/Y}, Jacobian \eqn{|J| = 1/Y^2}.
}
\examples{
d <- transformation(gamma_distrib(), inverse_transform())
distrib_pdf(d, 1, list(mu = 2, sigma2 = 1))

}
