% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distInverseGamma.R
\name{InverseGamma}
\alias{InverseGamma}
\title{Inverse gamma distribution}
\usage{
InverseGamma(alpha, theta, lambda, mu)
}
\arguments{
\item{alpha}{shape parameter.}

\item{theta}{scale parameter.}

\item{lambda}{rate parameter.}

\item{mu}{mean parameter.}
}
\description{
Create an inverse gamma distribution object.
}
\note{
Parameter meanings are on the inverse scale, e.g., \code{mu} is the mean of 1/x.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
