% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis.Gamma1Distrib}
\alias{kurtosis.Gamma1Distrib}
\title{Kurtosis of the Gamma in Mean and Dispersion}
\arguments{
\item{x}{A \code{Gamma1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: excess \eqn{6\phi}, free of the mean.
}
\keyword{internal}
