% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{logit_transform}
\alias{logit_transform}
\title{Logit Transformation}
\usage{
logit_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \text{logit}(X)}: maps \eqn{(0, 1)} to the real line.
Inverse \eqn{X = \text{plogis}(Y)}, Jacobian \eqn{|J| = \text{dlogis}(Y)}.
}
