% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{logit_transform}
\alias{logit_transform}
\title{Logit Transformation}
\usage{
logit_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \text{logit}(X)}: maps \eqn{(0, 1)} to the real line.
Inverse \eqn{X = \text{plogis}(Y)}, Jacobian \eqn{|J| = \text{dlogis}(Y)}.
}
\examples{
d <- transformation(beta1_distrib(), logit_transform())
distrib_pdf(d, 0, list(mu = 0.4, phi = 5))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{bc_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
