% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distrib.R
\name{logLik.distrib_fit}
\alias{logLik.distrib_fit}
\title{Log-Likelihood of a Maximum-Likelihood Fit}
\arguments{
\item{object}{A \code{\link[=distrib_fit]{distrib_fit()}} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{logLik}.
}
\description{
Log-Likelihood of a Maximum-Likelihood Fit
}
