% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.BinomialDistrib}
\alias{mean.BinomialDistrib}
\title{Mean of the Binomial Distribution}
\arguments{
\item{x}{A \code{BinomialDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{n\mu}.
}
\keyword{internal}
