% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_distrib.R
\name{mean.DirichletDistrib}
\alias{mean.DirichletDistrib}
\title{Mean of a Dirichlet}
\arguments{
\item{x}{A \code{DirichletDistrib} object.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector summing to one.
}
\description{
The mean vector, which is a parameter of the family.
}
\seealso{
\code{\link[=dirichlet_distrib]{dirichlet_distrib()}}
}
\keyword{internal}
