% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.Gaussian1Distrib}
\alias{mean.Gaussian1Distrib}
\title{Mean of the Gaussian Distribution}
\arguments{
\item{x}{A \code{Gaussian1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{E[Y] = \mu}.
}
\keyword{internal}
