% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distrib.R
\name{mean.MultinomialDistrib}
\alias{mean.MultinomialDistrib}
\title{Mean of a Multinomial}
\arguments{
\item{x}{A \code{MultinomialDistrib} object.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector summing to the size.
}
\description{
The vector \eqn{np}, which \code{\link[=mv_location]{mv_location()}} returns.
}
\seealso{
\code{\link[=multinomial_distrib]{multinomial_distrib()}}
}
\keyword{internal}
