% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgaussian_distrib.R
\name{mean.MvGaussianDistrib}
\alias{mean.MvGaussianDistrib}
\title{Mean of a Multivariate Gaussian}
\arguments{
\item{x}{A \code{\link[=MvGaussianDistrib]{MvGaussianDistrib()}} object.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector of length \eqn{p}.
}
\description{
The mean vector, which is a parameter of the family.
}
\keyword{internal}
