% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.NegBin1Distrib}
\alias{mean.NegBin1Distrib}
\title{Mean of the NB1 Negative Binomial Distribution}
\arguments{
\item{x}{A \code{NegBin1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu}.
}
\keyword{internal}
