% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mean.PoissonDistrib}
\alias{mean.PoissonDistrib}
\title{Mean of the Poisson Distribution}
\arguments{
\item{x}{A \code{PoissonDistrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu}.
}
\keyword{internal}
