% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal2_distrib.R
\name{mean.SkewNormal2Distrib}
\alias{mean.SkewNormal2Distrib}
\title{Mean of the Skew Normal in the Centred Parametrisation}
\arguments{
\item{x}{A \code{SkewNormal2Distrib} object.}

\item{theta}{A list with \code{mu}, \code{sigma} and \code{gamma1}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\mu}, the parameter itself: that is what centred means.
}
\keyword{internal}
