% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgaussian_distrib.R
\name{mv_location.MvGaussianDistrib}
\alias{mv_location.MvGaussianDistrib}
\title{Mean of a Multivariate Gaussian}
\arguments{
\item{distrib}{A \code{\link{MvGaussianDistrib}} object.}

\item{theta}{A named list of parameters.}
}
\value{
A named numeric vector of length \eqn{p}.
}
\description{
The first \eqn{p} parameters, which are the mean vector.
}
\keyword{internal}
