% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated.R
\name{n_support_points}
\alias{n_support_points}
\title{Number of Points in a Discrete Support}
\usage{
n_support_points(distrib)
}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}
}
\value{
A single number.
}
\description{
How many points the distribution's support contains, \code{Inf} when it is
unbounded.
}
\seealso{
\code{\link[=check_support_is_rich_enough]{check_support_is_rich_enough()}}
}
\keyword{internal}
