% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNegativeBinomial.R
\name{NegativeBinomial}
\alias{NegativeBinomial}
\title{Negative binomial distribution}
\usage{
NegativeBinomial(mu, phi, p, k)
}
\arguments{
\item{mu}{mean parameter.}

\item{phi}{dispersion parameter.}

\item{p}{probability of success parameter.}

\item{k}{number of successes parameter.}
}
\description{
Create a negative binomial distribution object.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
