% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal.R
\name{normal}
\alias{normal}
\alias{NormalSigma}
\alias{NormalSigma2}
\alias{NormalTau}
\title{Normal distribution}
\usage{
normal(mu, sigma, sigma2, tau)

NormalSigma(mu, sigma)

NormalSigma2(mu, sigma2)

NormalTau(mu, tau)
}
\arguments{
\item{mu}{Mean.}

\item{sigma}{Standard deviation.}

\item{sigma2}{Variance.}

\item{tau}{Precision.}
}
\value{
Object of class [NormalSigma()], [NormalSigma2()], or [NormalTau()].
}
\description{
Create a normal distribution object. It is advised to use the factory method [normal()] rather than calling the
constructors of the indivitual classes.
}
