% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-estimation.R
\name{parameter-estimation}
\alias{parameter-estimation}
\alias{fit.Distribution}
\alias{parameter_estimates}
\alias{parameter_start}
\alias{parameter_inference}
\title{Parameter estimation}
\usage{
\method{fit}{Distribution}(object, estimator = Mle(), data, ...)

parameter_estimates(distribution, estimator, data)

parameter_start(distribution, data)

parameter_inference(distribution, inference_method, data)
}
\arguments{
\item{object}{Object of class \code{\link[=Distribution]{Distribution()}}.}

\item{estimator}{Object of class \code{\link[=Estimator]{Estimator()}}.}

\item{data}{numeric vector.}

\item{...}{For S3 compatibility; not used.}

\item{distribution}{Object of class \code{\link[=Distribution]{Distribution()}}.}

\item{inference_method}{Object of class \code{\link[=InferenceMethod]{InferenceMethod()}}.}
}
\value{
\code{\link[=parameter_estimates]{parameter_estimates()}} returns a named list with parameter values.
\code{\link[=parameter_inference]{parameter_inference()}} returns a data.frame with parameter estimates, optionally SE and CI or other summaries.
\code{\link[=parameter_start]{parameter_start()}} and \code{\link[=fit]{fit()}} return an object of class \code{\link[=Distribution]{Distribution()}}.
}
\description{
Methods for estimating parameters of a distribution.
}
\note{
\code{\link[=fit]{fit()}} is recommended if you want to fit a distribution to the data. In addition to estimating the parameters,
it also validates the final distribution object.
}
