% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{parameter-estimation}
\alias{parameter-estimation}
\alias{parameter_estimates}
\alias{parameter_start}
\alias{parameter_inference}
\alias{fit_distribution}
\title{Parameter estimation}
\usage{
parameter_estimates(distribution, estimator, data)

parameter_start(distribution, data)

parameter_inference(distribution, inference_method, data)

fit_distribution(distribution, estimator, data)
}
\arguments{
\item{distribution}{Object of class [Distribution()].}

\item{estimator}{Object of class [Estimator()].}

\item{data}{numeric vector.}

\item{inference_method}{Object of class [InferenceMethod()].}
}
\value{
[parameter_estimates()] returns a named list with parameter values.
[parameter_inference()] returns a data.frame with parameter estimates, optionally SE and CI or other summaries.
[parameter_start()] and [fit_distribution()] return an object of class [Distribution()].
}
\description{
Methods for estimating parameters of a distribution.
}
