% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{parameter-inference}
\alias{parameter-inference}
\alias{InferenceMethod}
\alias{DefaultMethod}
\alias{NormalTheory}
\alias{Bootstrap}
\title{Parameter inference methods}
\usage{
InferenceMethod(estimator = <object>, ci_level = 0.95)

DefaultMethod(estimator = <object>, ci_level = 0.95, args = list())

NormalTheory(
  estimator = <object>,
  ci_level = 0.95,
  constrained = FALSE,
  control = list()
)

Bootstrap(
  estimator = <object>,
  ci_level = 0.95,
  samples = 1000L,
  callback = function () 
 {
 }
)
}
\arguments{
\item{estimator}{Object of class [Estimator()].}

\item{ci_level}{numeric; confidence level.}

\item{args}{list; optional arguments for the method.}

\item{constrained}{logical; should the CIs be computed on the unconstrained space (and then transformed),
or should be computed directly on the constrained space.}

\item{control}{list; passed to [stats::optimHess].}

\item{samples}{integer; How many samples to take.}

\item{callback}{function; Optional function to exectute on every iteration.}
}
\description{
These classes encapsulate methods for parameter inference.
}
