% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-empirical.R
\name{plot-empirical}
\alias{plot-empirical}
\alias{plot_empirical}
\alias{plot_hist}
\alias{plot_qq}
\alias{plot_ecdf}
\alias{plot_pp}
\title{Empirical plots}
\usage{
plot_empirical(
  distribution,
  data,
  type = c("all", "hist", "qq", "ecdf", "pp"),
  ...,
  name,
  ci = FALSE,
  ci_level = 0.95
)

plot_hist(
  distribution,
  data,
  ...,
  name,
  bin_width_type = "doane",
  number_of_bins = NA
)

plot_qq(distribution, data, ..., ci = FALSE, ci_level = 0.95)

plot_ecdf(distribution, data, ..., name)

plot_pp(distribution, data, ..., ci = FALSE, ci_level = 0.95)
}
\arguments{
\item{distribution}{Object of class \code{\link[=Distribution]{Distribution()}}.}

\item{data}{Numeric containing the data.}

\item{type}{Character; which plots to plot.}

\item{...}{Not used.}

\item{name}{Character; Name of the variable.}

\item{ci}{Logical; Should confidence bands be plotted?}

\item{ci_level}{Numeric; Confidence level of confidence bands.}

\item{bin_width_type}{Character; Passed to \code{\link[jaspGraphs:jaspHistogram]{jaspGraphs::jaspHistogramBinWidth()}}.}

\item{number_of_bins}{Integer; Passed to \code{\link[jaspGraphs:jaspHistogram]{jaspGraphs::jaspHistogramBinWidth()}}.}
}
\description{
Plot a distribution against data.
}
