% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{power_transform}
\alias{power_transform}
\title{Power Transformation}
\usage{
power_transform(p = 2)
}
\arguments{
\item{p}{Numeric. The exponent. Defaults to 2.}
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = X^p}. Inverse \eqn{X = Y^{1/p}},
Jacobian \eqn{|J| = \frac{1}{|p|}|Y|^{1/p - 1}}. Fractional powers require a
non-negative support; even integer powers require a support that does not straddle 0.
}
\examples{
d <- transformation(gamma2_distrib(), power_transform(p = 2))
distrib_pdf(d, 1, list(mu = 2, sigma2 = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{bc_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{logit_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
