% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.distrib}
\alias{print.distrib}
\title{Print Method for \code{distrib} Objects}
\arguments{
\item{x}{An object inheriting from class \code{"distrib"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The object \code{x} invisibly.
}
\description{
Custom S7 print method for objects inheriting from \code{distrib}.
}
\keyword{internal}
