% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distrib.R
\name{print_check_table}
\alias{print_check_table}
\title{Print a Validation Table}
\usage{
print_check_table(distrib, out, theta, n, nsim)
}
\arguments{
\item{distrib}{An object inheriting from class \code{"distrib"}.}

\item{out}{The data frame of checks.}

\item{theta}{The parameters the checks were run at.}

\item{n}{The number of observations used.}

\item{nsim}{The Monte Carlo sample size used.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Renders what \code{\link[=check_distrib]{check_distrib()}} found, in the same shape for a
univariate and a multivariate distribution.
}
\seealso{
\code{\link[=check_distrib]{check_distrib()}}
}
\keyword{internal}
