% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-shifted.R, R/dist-shifted_exponential.R
\name{Shifted}
\alias{Shifted}
\alias{ShiftedLogNormal}
\alias{ShiftedGamma}
\alias{ShiftedInverseGamma}
\alias{ShiftedLogLogistic}
\alias{ShiftedWald}
\alias{ShiftedWeibull}
\alias{ShiftedExponential}
\title{Shift a distribution by a scalar.}
\usage{
Shifted(distribution, shift = 0)

ShiftedLogNormal(mu, sigma, shift = 0)

ShiftedGamma(alpha, theta, lambda, mu, shift = 0)

ShiftedInverseGamma(alpha, theta, lambda, mu, shift = 0)

ShiftedLogLogistic(mu, sigma, alpha, beta, shift = 0)

ShiftedWald(mu, lambda, nu, alpha, sigma = fixed(1), shift = 0)

ShiftedWeibull(shape, scale, shift = 0)

ShiftedExponential(lambda, beta, shift = 0)
}
\arguments{
\item{distribution}{An object of class \code{\link[=Distribution]{Distribution()}}.}

\item{shift}{A shift parameter.}

\item{mu}{log mean parameter.}

\item{sigma}{log standard deviation parameter.}

\item{alpha}{shape parameter.}

\item{theta}{scale parameter.}

\item{lambda}{rate parameter.}

\item{beta}{scale parameter.}

\item{nu}{drift rate parameter.}

\item{shape}{shape parameter.}

\item{scale}{scale parameter.}
}
\description{
Create a distribution that has an additional shift parameter.
}
\note{
Estimating the shift parameter might not be a good idea, depending on the distribution. Distributions with their
own location parameter (e.g., normal), do not need a shift parameter in the first place.

For convenience, additional distribution factories are added for distributions that are typical to parametrise with
an additional shift parameter. For some, custom fitting methods are used to estimate the shift parameter.
}
\seealso{
\code{\link[=LogNormal]{LogNormal()}}, \code{\link[=Gamma]{Gamma()}}, \code{\link[=InverseGamma]{InverseGamma()}}, \code{\link[=LogLogistic]{LogLogistic()}}, \code{\link[=Wald]{Wald()}}, \code{\link[=Weibull]{Weibull()}}, \code{\link[=Exponential]{Exponential()}}.

Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
