% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distSkew.R
\name{skew-distributions}
\alias{skew-distributions}
\alias{SkewNormal}
\alias{SkewCauchy}
\alias{SkewT}
\title{Skew distribution family}
\usage{
SkewNormal(xi, omega, alpha)

SkewCauchy(xi, omega, alpha)

SkewT(xi, omega, alpha, nu)
}
\arguments{
\item{xi}{location parameter.}

\item{omega}{scale parameter.}

\item{alpha}{slant (skew) parameter.}

\item{nu}{degrees of freedom parameter.}
}
\description{
Create an object of the Skew distribution family.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{t-distribution}}
}
\concept{distributions}
