% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewed_generalized_t.R
\name{SkewedGeneralizedT}
\alias{SkewedGeneralizedT}
\title{Skewed generalized t-distribution}
\usage{
SkewedGeneralizedT(mu, sigma, lambda, p, q)
}
\arguments{
\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{lambda}{skew parameter.}

\item{p}{kurtosis parameter.}

\item{q}{kurtosis parameter.}
}
\description{
Create a skewed generalized t-distribution.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
