% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.distrib}
\alias{skewness.distrib}
\title{Skewness of a Distribution}
\arguments{
\item{x}{A \code{distrib} object.}

\item{theta}{A named list of parameters.}

\item{...}{Passed to \code{\link{moment}}.}
}
\value{
A numeric vector.
}
\description{
Computes the standardised third central moment through \code{\link{moment}}.
}
\keyword{internal}
