% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness.InvGauss2Distrib}
\alias{skewness.InvGauss2Distrib}
\title{Skewness of the Inverse Gaussian in Mean and Shape}
\arguments{
\item{x}{An \code{InvGauss2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{3\sqrt{\mu/\lambda}}.
}
\keyword{internal}
