% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnormal2_distrib.R
\name{skewness.SkewNormal2Distrib}
\alias{skewness.SkewNormal2Distrib}
\title{Skewness of the Skew Normal in the Centered Parametrization}
\arguments{
\item{x}{A \code{SkewNormal2Distrib} object.}

\item{theta}{A list with \code{mu}, \code{sigma} and \code{gamma1}.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
\eqn{\gamma_1}, the parameter itself.
}
\keyword{internal}
