% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{sqrt_transform}
\alias{sqrt_transform}
\title{Square Root Transformation}
\usage{
sqrt_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \sqrt{X}} (requires \eqn{X \ge 0}).
Inverse \eqn{X = Y^2}, Jacobian \eqn{|J| = 2Y}.
}
