% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{sqrt_transform}
\alias{sqrt_transform}
\title{Square Root Transformation}
\usage{
sqrt_transform()
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for \eqn{Y = \sqrt{X}} (requires \eqn{X \ge 0}).
Inverse \eqn{X = Y^2}, Jacobian \eqn{|J| = 2Y}.
}
\examples{
d <- transformation(gamma2_distrib(), sqrt_transform())
distrib_pdf(d, 1, list(mu = 2, sigma2 = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{bc_transform}}, \code{\link{yj_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{logit_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
