% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{std_dev.distrib}
\alias{std_dev.distrib}
\title{Standard Deviation of a Distribution}
\arguments{
\item{x}{A \code{distrib} object.}

\item{theta}{A named list of parameters.}

\item{...}{Passed to \code{\link[=moment]{moment()}}.}
}
\value{
A numeric vector.
}
\description{
The square root of \code{\link[=variance]{variance()}}.
}
\keyword{internal}
