% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{std_dev}
\alias{std_dev}
\title{Standard Deviation of a Distribution or Sample}
\usage{
std_dev(x, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{"distrib"}, or a numeric vector.}

\item{...}{For \code{distrib} objects: \code{theta} and further arguments passed to
\code{\link{moment}}. For numeric vectors: \code{na.rm}.}
}
\value{
A numeric vector.
}
\description{
Computes the standard deviation as the square root of \code{\link{variance}}.
For numeric vectors the sample standard deviation \code{\link[stats]{sd}} is returned.
}
\examples{
std_dev(gaussian1_distrib(), list(mu = 0, sigma = 2))

}
