% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{support}
\alias{support}
\alias{Support}
\alias{Real}
\alias{Int}
\title{Support classes}
\usage{
Support(min = -Inf, max = Inf)

Real(min = -Inf, max = Inf)

Int(min = -Inf, max = Inf)
}
\arguments{
\item{min}{Minimum of the support. Can be either a numeric, or an expression.}

\item{max}{Maximum of the support. Can be either a numeric, or an expression.}
}
\description{
Classes the encapsulate support of distributions or parameters.
}
