% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-noncentral_student_t.R
\name{t-distribution}
\alias{t-distribution}
\alias{noncentral_student_t}
\alias{noncentral_t}
\alias{student_t}
\alias{StandardT}
\alias{StudentT}
\alias{NoncentralT}
\alias{NoncentralStudentT}
\title{T-distribution}
\usage{
noncentral_student_t(nu, kappa, mu, sigma)

noncentral_t(nu, kappa)

student_t(nu, mu, sigma)

StandardT(nu)

StudentT(nu, mu, sigma)

NoncentralT(nu, kappa, mu, sigma)

NoncentralStudentT(nu, kappa, mu, sigma)
}
\arguments{
\item{nu}{degrees of freedom parameter.}

\item{kappa}{noncentrality parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}
}
\description{
Create a t-distribution object.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Wald}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}}
}
\concept{distributions}
