% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{trunc_hessian}
\alias{trunc_hessian}
\title{Observed Hessian of a Truncated Distribution}
\usage{
trunc_hessian(distrib, y, theta, scale = c("parameter", "link"), ...)
}
\arguments{
\item{distrib}{A truncated distribution object.}

\item{y}{A numeric vector of observations.}

\item{theta}{A named list of parameters.}
}
\value{
A named list, one component per Hessian entry.
}
\description{
\eqn{d_{ij}\ell_T = H_{ij}(y) - M_{ij} + m_i m_j}.
}
\seealso{
\code{\link{trunc_M}}, \code{\link{trunc_score_mean}}
}
\keyword{internal}
