% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{trunc_inside}
\alias{trunc_inside}
\title{Which Observations Lie in the Truncated Support}
\usage{
trunc_inside(distrib, y)
}
\arguments{
\item{distrib}{A truncated distribution object.}

\item{y}{A numeric vector of observations.}
}
\value{
A logical vector as long as \code{y}.
}
\description{
Which Observations Lie in the Truncated Support
}
\keyword{internal}
