% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.Gaussian1Distrib}
\alias{variance.Gaussian1Distrib}
\title{Variance of the Gaussian Distribution}
\arguments{
\item{x}{A \code{Gaussian1Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\operatorname{Var}(Y) = \sigma^2}.
}
\keyword{internal}
