% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.GumbelDistrib}
\alias{variance.GumbelDistrib}
\title{Variance of the Gumbel Distribution}
\arguments{
\item{x}{A \code{\link[=GumbelDistrib]{GumbelDistrib()}}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\pi^2\sigma^2/6}.
}
\keyword{internal}
