% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.InvGauss2Distrib}
\alias{variance.InvGauss2Distrib}
\title{Variance of the Inverse Gaussian in Mean and Shape}
\arguments{
\item{x}{An \code{InvGauss2Distrib}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\mu^3/\lambda}.
}
\keyword{internal}
