% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{variance.SkewNormal1Distrib}
\alias{variance.SkewNormal1Distrib}
\title{Variance of the Skew Normal Distribution}
\arguments{
\item{x}{A \code{\link[=SkewNormal1Distrib]{SkewNormal1Distrib()}}.}

\item{theta}{A named list of parameters.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Closed form: \eqn{\sigma^2\left(1 - b^2\delta^2\right)}.
}
\keyword{internal}
