% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distrib.R
\name{vcov.distrib_fit}
\alias{vcov.distrib_fit}
\title{Variance-Covariance Matrix of a Maximum-Likelihood Fit}
\arguments{
\item{object}{A \code{\link[=distrib_fit]{distrib_fit()}} object.}

\item{scale}{Either \code{"parameter"} (default) or \code{"link"}.}

\item{...}{Unused.}
}
\value{
A variance-covariance matrix.
}
\description{
Variance-Covariance Matrix of a Maximum-Likelihood Fit
}
