% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-wald.R
\name{Wald}
\alias{Wald}
\title{Wald (inverse gaussian) distribution}
\usage{
Wald(mu, lambda, nu, alpha, sigma = fixed(1))
}
\arguments{
\item{mu}{mean parameter.}

\item{lambda}{shape parameter.}

\item{nu}{drift rate parameter.}

\item{alpha}{threshold parameter.}

\item{sigma}{noise parameter.}
}
\description{
Create a Wald distribution object.
}
\seealso{
Other distributions: 
\code{\link{Amoroso}()},
\code{\link{Beta}()},
\code{\link{BetaPrime}()},
\code{\link{Binomial}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquared}()},
\code{\link{Exponential}()},
\code{\link{Frechet}()},
\code{\link{Gamma}()},
\code{\link{Gompertz}()},
\code{\link{Gumbel}()},
\code{\link{InverseGamma}()},
\code{\link{Laplace}()},
\code{\link{LogLogistic}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{NegativeBinomial}()},
\code{\link{Normal}()},
\code{\link{Pareto}()},
\code{\link{Shifted}()},
\code{\link{SkewedGeneralizedT}()},
\code{\link{StretchedBeta}()},
\code{\link{SymmetricGeneralizedNormal}()},
\code{\link{Triangular}()},
\code{\link{Uniform}()},
\code{\link{Weibull}()},
\code{\link{distributions}},
\code{\link{f-distribution}},
\code{\link{skew-distributions}},
\code{\link{t-distribution}}
}
\concept{distributions}
