% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{yj_transform}
\alias{yj_transform}
\title{Yeo-Johnson Transformation}
\usage{
yj_transform(lambda)
}
\arguments{
\item{lambda}{Numeric. The transformation parameter.}
}
\value{
A \code{\link{transformer}} object.
}
\description{
Transformer for the Yeo-Johnson transformation, extending Box-Cox to
negative values: Box-Cox of \eqn{X+1} (parameter \eqn{\lambda}) for \eqn{X \ge 0},
and negated Box-Cox of \eqn{|X|+1} (parameter \eqn{2-\lambda}) for \eqn{X < 0}.
}
\examples{
d <- transformation(gaussian1_distrib(), yj_transform(lambda = 0.5))
distrib_pdf(d, 1, list(mu = 0, sigma = 1))

}
\seealso{
\code{\link{log_transform}}, \code{\link{exp_transform}}, \code{\link{sqrt_transform}}, \code{\link{inverse_transform}}, \code{\link{power_transform}}, \code{\link{bc_transform}}, \code{\link{softplus_transform}}, \code{\link{asinh_transform}}, \code{\link{logit_transform}}, \code{\link{expit_transform}}, \code{\link{affine_transform}}, \code{\link{transformation}}
}
